### 2026-02-20 #285 @nstarman
`phasecurvefit`
Submitting Author: (@nstarman)
All current maintainers: (@nstarman, @jnibauer)
Package Name: phasecurvefit
Construct paths through phase-space points, with JAX.
Repository Link:  https://github.com/GalacticDynamics/phasecurvefit
Version submitted: v0.1.0
Version accepted: TBD

---

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in future interactions in spaces supported by pyOpenSci should it be accepted.

Filamentary structures are ubiquitous in the physical sciences, ranging from coherent streams of stars called stellar streams to elongated structures in turbulent fluids, plasmas, and the interstellar medium. In the context of stellar streams, a common preprocessing step is ordering observational or simulation data to derive the mean path of the stream's trajectory through phase-space.

`phasecurvefit` is an open-source Python package for constructing such orderings and paths by _walking along the local phase-space flow_ using. There are two core components: the first is `walk_local_flow` which builds an approximate ordering and trajectory but which might miss some of the data; and the second is a fast-to-train autoencoder that imputes the full ordering and trajectory.

If your package is associated with an  existing community please check below:

**While this package does not directly interface with Astropy, it easily interoperates with it, through [`unxt`](https://pypi.org/project/unxt/), which brings Astropy to JAX.**

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

The target audience is scientists with data or simulations of stellar streams, or similar filamentary structures like elongated structures in turbulent fluids, plasmas, and the interstellar medium. Current scientific applications of this package, not by the submitters, are analyses of Milky Way stellar streams data, fitting the path of the stream for orbit and membership likelihood analyses.

Not to our knowledge. Our code, written in JAX, not only fits stream-like structures in phase space, it does so differentiably e.g. for use in gradient-utilizing inference pipelines.

- [x] The package contains a `paper.md` matching [JOSS's requirements][JossPaperRequirements] with a high-level description in the package root or in `inst/`. **Yes, in the PR https://github.com/GalacticDynamics/phasecurvefit/pull/3, which will be merged on acceptance**
- [x] The package is deposited in a long-term repository with the DOI: [10.5281/zenodo.18714341](https://doi.org/10.5281/zenodo.18714341)
