### 2026-02-25 GalacticDynamics/phasecurvefit#15 @crhea93
Error when importing phasecurve
Hello,
 I've run into an error (likely with a  dependency) when trying to import the module in a new environment. I installed the module following the instructions in the documentation (i.e. `pip install phasecurvefit`). However, I receive and error when I try to import it. Below you will find the error:

```
Python 3.13.11 | packaged by Anaconda, Inc. | (main, Dec 10 2025, 21:28:48) [GCC 14.3.0] on linux
>>> import phasecurvefit as pcf
...
Traceback (most recent call last):
  File "<python-input-0>", line 1, in <module>
    import phasecurvefit as pcf
  File "/home/carterrhea/miniconda3/lib/python3.13/site-packages/phasecurvefit/__init__.py", line 98, in <module>
    from . import _interop  # noqa: F401
    ^^^^^^^^^^^^^^^^^^^^^^
  File "/home/carterrhea/miniconda3/lib/python3.13/site-packages/phasecurvefit/_interop/__init__.py", line 18, in <module>
    from . import interop_unxt  # noqa: F401
    ^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/carterrhea/miniconda3/lib/python3.13/site-packages/phasecurvefit/_interop/interop_unxt.py", line 42, in <module>
    import unxt as u
  File "/home/carterrhea/miniconda3/lib/python3.13/site-packages/unxt/__init__.py", line 56, in <module>
    from . import dims, quantity, units, unitsystems
  File "/home/carterrhea/miniconda3/lib/python3.13/site-packages/unxt/quantity.py", line 158, in <module>
    from ._src.quantity import (
    ...<4 lines>...
    )
  File "/home/carterrhea/miniconda3/lib/python3.13/site-packages/unxt/_src/quantity/register_dispatches.py", line 12, in <module>
    from plum.parametric import type_unparametrized as type_np
ModuleNotFoundError: No module named 'plum.parametric'
```

This issue is related to the PyOpenSci submission https://github.com/pyOpenSci/software-submission/issues/285
