### 2026-02-27 f16092b
Backport PR #16: 👽️ external: jaxmore (#17)
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
