### 2026-02-27 #285 @nstarman
> Thank you for this submission. It is well conceived demonstrating a substantial effort on your part. There are however a few things that we would like to see addressed before assigning an editor.

Thanks! Happy to address

> PyOpenSci is committed to ensuring that all code developed is readily maintainable in light of recent agentic coding. I notice that a high number of your commits seem to be generated by generative AI. While this is certainly not a problem in-and-of itself, it is important that you can vouch for every line of code created by AI on your behalf. Therefore, we require a detailed report in the README disclosing the use of generative AI.

I'd like to push back on this point in two ways. The first is that the commits are "generated" by AI. I reviewed each commit (13) and can confidently state that the vast majority of the operative code was written by me. Adapting, improving, and fine-tuning the mathematical techniques. Crafting the API. That was not AI. AI was used as a code assistant for writing the docs(strings), type annotations, and unit tests. As is the case with generated content, significant editing was required. Especially since the docs are tested as part of CI. So I don't believe either a high number of commits nor most of the core code was AI generated, just the surrounding supports. The infrastructure of this package is closely copied from the infrastructure I have been developing across my many other packages: for example the 22 repositories I maintain at https://github.com/orgs/GalacticDynamics/repositories. This speaks to the challenging issue of determining authorship, when for example I ask an AI agent to "copy the infrastructure from https://github.com/orgs/GalacticDynamics/unxt and adapt it to this code repository". Is this generated by AI or is this mostly written by me (in `unxt`) and an AI changed the word `unxt` to `phasecurvefit`? I would like to think that having spent significant time developing the shared architecture between the GalacticDynamics org's repos, that copy-paste-with-small changes wouldn't erase my authorship.
Either way, I agree that taking responsibility for one's code is important. To this end I have a PR adding to the Contributing.md, adapting the AI policy we're still developing for Astropy — https://github.com/GalacticDynamics/phasecurvefit/pull/18.
The second point — which might not be relevant given the first point — is that the README is the appropriate place for a disclosure report on who/what authored which lines. Generative AI runs the gamut from line-completions to file-completions. And the README is a living document that is also published to PyPI. As an active code that will continue to see development, such a disclosure is guaranteed to be quickly outdated and should be removed essentially immediately as it's relevant only to a specific moment.
If required I would produce a separate document relevant to this review process, but as the policy in https://github.com/GalacticDynamics/phasecurvefit/pull/18 indicates, I would more happily "explain all [code] during review" as I wrote most of it and reviewed, tested, and understand all of it.

Thank you!
