### 2026-02-28 #285 @crhea93
Hi Nathan,
 My apologies if my original comments were not clear enough! I didn't mean to insinuate (and I sincerely apologize if I did) that the code itself was written by AI and not you. I  meant that the commit messages are often signed by Co-pilot and the text of the commit is very AI-coded (emojis & language). There is no worries at all if gen AI was used to help build the support, documentation, or tests.

We believe that the README is the perfect place for the disclosure of use of generative AI since it is the first place people look when they  check out a repo. We don't expect the authors to have a detailed description of each location in which they used generative AI in the README since, as you mention, that could change drastically over the life of a project. However, we do believe that a README can be reasonably written and updated to include the major AI usage.
This repo has an example of a minimal acceptable gen AI disclosure : https://github.com/DynamicsAndNeuralSystems/pyhctsa
