### 2026-03-02 6979602
📝 docs: update readme
with AI disclosure statement, adapted from https://github.com/DynamicsAndNeuralSystems/pyhctsa

Signed-off-by: nstarman <nstarman@users.noreply.github.com>
