### 2026-03-02 dd040de
Backport PR #22: 📝 docs: update readme
