+++ README.md
@@ -6,6 +6,21 @@
 Construct paths through phase-Space points, supporting many different
 algorithms.
 
+## Features
+
+- **JAX-powered**: Fully compatible with JAX transformations (`jit`, `vmap`,
+  `grad`)
+- **GPU-ready**: Runs on CPU, GPU, or TPU via JAX
+- **Type-safe**: Comprehensive (optionally runtime checked) type hints with
+  `jaxtyping`
+- **Pluggable metrics**: Customizable distance metrics for different physical
+  interpretations
+- **Pluggable query strategies**: Flexible neighbor search strategies (e.g.,
+  brute-force, KD-tree) to optimize performance
+- **Highly customizable ML setup and training**: Well-chosen defaults with
+  highly flexible customization for specific use-cases.
+- **Physical units**: Optional support via `unxt` for unit-aware calculations
+
 ## Installation
 
 Install the core package:
@@ -66,66 +81,106 @@ uv add phasecurvefit --extra kdtree
 ## Quick Start
 
 ```python
+import jax
 import jax.numpy as jnp
 import phasecurvefit as pcf
 
 # Create phase-space observations as dictionaries (Cartesian coordinates)
-pos = {"x": jnp.array([0.0, 1.0, 2.0]), "y": jnp.array([0.0, 0.5, 1.0])}
-vel = {"x": jnp.array([1.0, 1.0, 1.0]), "y": jnp.array([0.5, 0.5, 0.5])}
+pos = {
+    "x": jnp.array([0.0, 1.0, 2.0, 3.0, 4.0]),
+    "y": jnp.array([0.0, 0.5, 1.0, 1.5, 2.0]),
+}
+vel = {
+    "x": jnp.array([1.0, 1.0, 1.0, 1.0, 1.0]),
+    "y": jnp.array([0.5, 0.5, 0.5, 0.5, 0.5]),
+}
 
-# Order the observations (use KD-tree for spatial neighbor prefiltering)
-config = pcf.WalkConfig(strategy=pcf.strats.KDTree(k=2))
+# Step 1: Order the observations (use KD-tree for spatial neighbor prefiltering)
+config = pcf.WalkConfig(strategy=pcf.strats.KDTree(k=3))  # k=3 for this small dataset
 result = pcf.walk_local_flow(pos, vel, config=config, start_idx=0, metric_scale=1.0)
-print(result.indices)  # Array([0, 1, 2], dtype=int32)
+print(result.indices)  # Initial ordering
+
+# Step 2: Create normalizer and autoencoder
+key = jax.random.key(0)
+normalizer = pcf.nn.StandardScalerNormalizer(pos, vel)
+autoencoder = pcf.nn.PathAutoencoder.make(
+    normalizer, gamma_range=result.gamma_range, key=key
+)
+
+# Step 3: Configure and run training
+train_config = pcf.nn.TrainingConfig(
+    n_epochs_encoder=100,  # Encoder-only epochs
+    n_epochs_both=50,  # Joint training epochs
+    show_pbar=False,  # Disable progress bar
+)
+
+# Train the autoencoder
+result, _, losses = pcf.nn.train_autoencoder(
+    autoencoder, result, config=train_config, key=key
+)
+
+print(result.indices)  # Post-training ordering
 ```
 
 ### With Physical Units
 
-When `unxt` is installed, you can use physical units:
+When `unxt` is installed, you can use physical units throughout the workflow:
 
 ```python
+import jax
+import jax.numpy as jnp
+import phasecurvefit as pcf
 import unxt as u
 
 # Create phase-space observations with units
[…]
