### 2026-03-02 290d1e5
📝 docs: update readme (#22)
with AI disclosure statement, adapted from https://github.com/DynamicsAndNeuralSystems/pyhctsa

Signed-off-by: nstarman <nstarman@users.noreply.github.com>
