### 2026-03-02 #285 @nstarman
> The commit messages are often signed by Co-pilot

Oh. That's copilot review, for reviewing the PRs to find bugs / edge cases.

> commit is very AI-coded (emojis & language)

That's https://pypi.org/project/commitizen-gitmoji/ to enable me to standardize my construction of commit messages via the command line. Standard construction means [labeler](https://github.com/GalacticDynamics/phasecurvefit/blob/main/.github/labeler.yml) can auto-label PRs and [release](https://github.com/GalacticDynamics/phasecurvefit/blob/main/.github/release.yml) can auto-sort and categorize PRs into a release changelog. I recommend some form of https://www.conventionalcommits.org/en/v1.0.0/ to any project, I just prefer the gitmoji variant for visual aesthetics.

> README can be reasonably written and updated to include the major AI usage.

SGTM. https://github.com/GalacticDynamics/phasecurvefit/pull/23 adapts https://github.com/DynamicsAndNeuralSystems/pyhctsa .
