### 2026-03-02 8fc03ff
Backport PR #22: 📝 docs: update readme (#23)
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
