+++ src/phasecurvefit/strats.py
@@ -4,10 +4,7 @@ __all__: tuple[str, ...] = (
     "AbstractQueryStrategy",
     "BruteForce",
     "KDTree",
+    "QueryResult",
 )
 
-from ._src.strategies import (
-    AbstractQueryStrategy,
-    BruteForce,
-    KDTree,
-)
+from ._src.strategies import AbstractQueryStrategy, BruteForce, KDTree, QueryResult
