### 2026-03-02 #285 @crhea93
@nstarman Thank you very much for addressing those concerns! I really appreciate the clear responses :)

We will be proceeding with this submission!
