### 2026-03-04 482da3f
♻️ refactor: ScanMLP -> eqx.nn.MLP(scan=True)
Equinox's implementation is based off of the one from this library.

Signed-off-by: nstarman <nstarman@users.noreply.github.com>
