+++ pyproject.toml
@@ -314,12 +314,12 @@ filterwarnings = [
   "error",
   # pytest
   "ignore:The hookimpl.*uses old-style configuration options:pytest.PytestDeprecationWarning",
-  # jax
-  "ignore:.*pp_eqn_rules is deprecated:DeprecationWarning",
-  "ignore:.*Primitive is deprecated:DeprecationWarning",
-  "ignore:.*mapped_aval is deprecated:DeprecationWarning",
-  "ignore:.*unmapped_aval is deprecated:DeprecationWarning",
-  "ignore:.*NotMapped is deprecated:DeprecationWarning",
+  # jax (restrict to relevant modules: equinox and jax use these deprecated APIs)
+  "ignore:jax\\.core\\.pp_eqn_rules is deprecated:DeprecationWarning:(jax|equinox)(\\.|_)",
+  "ignore:jax\\._src\\.interpreters\\.mlir\\.Primitive is deprecated:DeprecationWarning:(jax|equinox)(\\.|_)",
+  "ignore:jax\\.core\\.mapped_aval is deprecated:DeprecationWarning:(jax|equinox)(\\.|_)",
+  "ignore:jax\\.core\\.unmapped_aval is deprecated:DeprecationWarning:(jax|equinox)(\\.|_)",
+  "ignore:jax\\.core\\.NotMapped is deprecated:DeprecationWarning:(jax|equinox)(\\.|_)",
   "ignore:The numpy\\.array_function protocol:DeprecationWarning",
   "ignore:numpy\\.ndarray size changed:RuntimeWarning",
   # matplotlib
