### 2026-03-04 44f00cd
⬆️ dep-bump: equinox v0.13.5+ (#24)
* ⬆️ dep-bump: equinox v0.13.5+
* ♻️ refactor: ScanMLP -> eqx.nn.MLP(scan=True). Equinox's implementation is based off of the one from this library.

Signed-off-by: nstarman <nstarman@users.noreply.github.com>
