### 2026-03-30 #285 @crhea93
Hi @nstarman,
I just wanted to give you a brief update. We are still looking for an editor -- at the moment we are starting the process to recruit more editors since the current number of editors is not quite sufficient for the amount of submissions we've been receiving recently :D
