### 2026-05-05 895f953
ci: taplo offline schema
Signed-off-by: nstarman <nstarman@users.noreply.github.com>
