+++ .pre-commit-config.yaml
@@ -91,4 +91,3 @@ repos:
     rev: v0.9.3
     hooks:
       - id: taplo-format
-      - id: taplo-lint
