### 2026-06-18 aa6d9ea
Set dependabot update check to maximum interval (yearly)
