### 2026-06-18 ceda036
fix: use valid Dependabot schedule interval (monthly)
