### 2026-07-02 5fc5865
✨ feat: pluggable orderer abstraction with unified result
Introduce `pcf.orderers` with `AbstractOrderer` (uniform host-side
`order()` interface), `LocalFlowOrderer` (wraps `walk_local_flow`,
regression-equivalent), and a `pcf.order` façade. Unify the result type as
`OrderingResult` (optional `backbone`, backbone-aware `__call__`);
`WalkLocalFlowResult` becomes a thin subclass, preserving its public name.
Delete the dead `abstract_walk_result.py`.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
