+++ tests/test_orderers.py
@@ -1,0 +1,114 @@
+"""Tests for the pluggable orderer abstraction (``pcf.orderers``)."""
+
+import jax.numpy as jnp
+import pytest
+
+import phasecurvefit as pcf
+from phasecurvefit._src.abstract_result import AbstractResult
+
+
+class TestOrdererNamespace:
+    """The public ``pcf.orderers`` surface."""
+
+    def test_exports(self):
+        """Exports."""
+        assert hasattr(pcf, "orderers")
+        for name in ("AbstractOrderer", "LocalFlowOrderer", "OrderingResult"):
+            assert hasattr(pcf.orderers, name), name
+
+    def test_order_facade_exists(self):
+        """Order facade exists."""
+        assert callable(pcf.order)
+
+
+class TestOrderingResultUnification:
+    """OrderingResult is unified; WalkLocalFlowResult subclasses it."""
+
+    def test_walklocalflowresult_subclasses_orderingresult(self):
+        """Walklocalflowresult subclasses orderingresult."""
+        assert issubclass(pcf.WalkLocalFlowResult, pcf.orderers.OrderingResult)
+        assert issubclass(pcf.orderers.OrderingResult, AbstractResult)
+
+    def test_backbone_aware_call_interpolates_along_backbone(self):
+        """Backbone aware call interpolates along backbone."""
+        # A straight backbone from (0,0) to (1,0), sampled at the two tips only.
+        positions = {"x": jnp.array([0.0, 0.5, 1.0]), "y": jnp.array([0.0, 0.0, 0.0])}
+        velocities = {"x": jnp.zeros(3), "y": jnp.zeros(3)}
+        backbone = {"x": jnp.array([0.0, 1.0]), "y": jnp.array([0.0, 0.0])}
+        res = pcf.orderers.OrderingResult(
+            positions=positions,
+            velocities=velocities,
+            indices=jnp.array([0, 1, 2]),
+            gamma_range=(-1.0, 1.0),
+            backbone=backbone,
+        )
+        mid = res(jnp.array(0.0))  # midpoint of gamma_range -> midpoint of backbone
+        assert jnp.allclose(mid["x"], 0.5)
+        assert jnp.allclose(mid["y"], 0.0)
+
+    def test_no_backbone_falls_back_to_ordered_points(self):
+        """No backbone falls back to ordered points."""
+        # Without a backbone, __call__ matches the legacy walk interpolation.
+        positions = {"x": jnp.array([0.0, 1.0, 2.0])}
+        velocities = {"x": jnp.zeros(3)}
+        res = pcf.orderers.OrderingResult(
+            positions=positions,
+            velocities=velocities,
+            indices=jnp.array([0, 1, 2]),
+        )
+        out = res(jnp.array(0.5))
+        assert jnp.allclose(out["x"], 1.0)
+
+
+class TestLocalFlowOrdererRegression:
+    """LocalFlowOrderer.order reproduces walk_local_flow bit-for-bit."""
+
+    @pytest.fixture
+    def data(self):
+        """Return sample phase-space data."""
+        q = {"x": jnp.array([0.0, 1.0, 2.0, 3.0, 4.0])}
+        p = {"x": jnp.array([1.0, 1.0, 1.0, 1.0, 1.0])}
+        return q, p
+
+    @pytest.mark.parametrize("direction", ["forward", "backward", "both"])
+    def test_matches_walk_local_flow(self, data, direction):
+        """Matches walk local flow."""
+        q, p = data
+        direct = pcf.walk_local_flow(
+            q, p, start_idx=0, metric_scale=1.0, direction=direction
+        )
+        orderer = pcf.orderers.LocalFlowOrderer(
[…]
