+++ src/phasecurvefit/_src/orderers/localflow.py
@@ -1,0 +1,78 @@
+"""The local-flow orderer: a backward-compatible wrapper around the walk."""
+
+__all__: tuple[str, ...] = ("LocalFlowOrderer",)
+
+import equinox as eqx
+import jax.numpy as jnp
+import plum
+
+from .base import AbstractOrderer
+from phasecurvefit._src.algorithm import (
+    Direction,
+    StateMetadata,
+    WalkLocalFlowResult,
+    walk_local_flow,
+)
+from phasecurvefit._src.custom_types import VectorComponents
+from phasecurvefit._src.query_config import WalkConfig
+
+
+class LocalFlowOrderer(AbstractOrderer):
+    """Order tracers with the velocity-following local-flow walk.
+
+    A thin wrapper around :func:`~phasecurvefit.walk_local_flow` so the walk is
+    usable through the uniform orderer interface. ``order()`` reproduces
+    ``walk_local_flow`` exactly (including ``direction="both"``, which the walk
+    stitches internally via ``combine_results``).
+
+    Parameters
+    ----------
+    metric_scale
+        Metric-dependent scale parameter (see ``walk_local_flow``).
+    config
+        Neighbor-query configuration (metric + strategy).
+    start_idx
+        Index of the starting observation.
+    direction
+        ``"forward"``, ``"backward"``, or ``"both"``.
+    max_dist
+        Maximum allowed neighbor distance.
+    terminate_indices
+        Indices that terminate the walk when reached.
+    n_max
+        Maximum number of iterations.
+
+    """
+
+    metric_scale: float = 1.0
+    config: WalkConfig = eqx.field(default_factory=WalkConfig)
+    start_idx: int = eqx.field(static=True, default=0)
+    direction: Direction = eqx.field(static=True, default="forward")
+    max_dist: float = jnp.inf
+    terminate_indices: frozenset[int] | None = eqx.field(static=True, default=None)
+    n_max: int | None = eqx.field(static=True, default=None)
+
+    @plum.dispatch
+    def order(
+        self,
+        positions: VectorComponents,
+        velocities: VectorComponents,
+        *,
+        metadata: StateMetadata | None = None,
+    ) -> WalkLocalFlowResult:
+        """Run the local-flow walk and return its result."""
+        kwargs: dict[str, object] = {}
+        if metadata is not None:
+            kwargs["metadata"] = metadata
+        return walk_local_flow(
+            positions,
+            velocities,
+            start_idx=self.start_idx,
+            metric_scale=self.metric_scale,
+            max_dist=self.max_dist,
+            terminate_indices=self.terminate_indices,
+            n_max=self.n_max,
+            config=self.config,
+            direction=self.direction,
+            **kwargs,
+        )
