### 2026-07-02 adc1907
✨ feat: MSTOrderer backbone ordering (spatial)
Add `MSTOrderer`: a host-side (NumPy/SciPy) orderer that builds a kNN graph,
takes its minimum spanning tree, finds the longest path (graph diameter) as a
tip-to-tip backbone, and orders every tracer by along-backbone arc length.
Needs no progenitor or start index. Handles disconnected graphs via
`on_disconnected` (raise/warn/largest). Adds scipy as a dependency.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
