+++ src/phasecurvefit/orderers.py
@@ -14,6 +14,9 @@ Built-in orderers
 -----------------
 LocalFlowOrderer
     Velocity-following greedy walk (wraps :func:`~phasecurvefit.walk_local_flow`).
+MSTOrderer
+    MST longest-path backbone ordering for near-closed-loop / self-overlapping
+    streams. Host-side (NumPy/SciPy).
 
 See Also
 --------
@@ -24,9 +27,11 @@ phasecurvefit.order : functional façade mirroring ``walk_local_flow``.
 __all__: tuple[str, ...] = (
     "AbstractOrderer",
     "LocalFlowOrderer",
+    "MSTOrderer",
     "OrderingResult",
 )
 
 from ._src.orderers.base import AbstractOrderer
 from ._src.orderers.localflow import LocalFlowOrderer
+from ._src.orderers.mst import MSTOrderer
 from ._src.orderers.result import OrderingResult
