+++ tests/test_mst.py
@@ -1,0 +1,160 @@
+"""Tests for the MST-backbone orderer (``pcf.orderers.MSTOrderer``)."""
+
+import jax.numpy as jnp
+import numpy as np
+import pytest
+
+import phasecurvefit as pcf
+from phasecurvefit._src.abstract_result import AbstractResult
+
+
+def _open_arc(n=200, seed=0):
+    """Points sampled along a known open 1-D curve, then shuffled.
+
+    Evenly spaced along the true parameter so the kNN graph stays connected
+    (random-uniform sampling can leave gaps wider than ``k`` can bridge).
+    """
+    rng = np.random.default_rng(seed)
+    t = np.linspace(0.0, 1.0, n)
+    x = 10.0 * t
+    y = np.sin(3.0 * t)
+    noise = rng.normal(0.0, 0.02, size=(n, 2))
+    pos = {"x": jnp.asarray(x + noise[:, 0]), "y": jnp.asarray(y + noise[:, 1])}
+    vel = {"x": jnp.ones(n), "y": jnp.asarray(3.0 * np.cos(3.0 * t))}
+    # shuffle so input order carries no ordering information
+    perm = rng.permutation(n)
+    pos = {k: v[perm] for k, v in pos.items()}
+    vel = {k: v[perm] for k, v in vel.items()}
+    t_shuffled = t[perm]
+    return pos, vel, jnp.asarray(t_shuffled)
+
+
+def _open_ring(n=240, gap_deg=20.0, seed=0):
+    """Points on a near-closed ring with a small angular gap (an open loop).
+
+    Evenly spaced in angle so the along-ring kNN graph stays connected while the
+    gap chord exceeds ``jump_cap`` and is severed.
+    """
+    rng = np.random.default_rng(seed)
+    ang = np.linspace(0.0, (360.0 - gap_deg), n) * np.pi / 180.0
+    x, y = np.cos(ang), np.sin(ang)
+    perm = rng.permutation(n)
+    pos = {"x": jnp.asarray(x[perm]), "y": jnp.asarray(y[perm])}
+    vel = {"x": jnp.asarray(-np.sin(ang)[perm]), "y": jnp.asarray(np.cos(ang)[perm])}
+    return pos, vel, jnp.asarray(ang[perm])
+
+
+class TestMSTNamespace:
+    """Tests for MST namespace."""
+
+    def test_exported(self):
+        """Exported."""
+        assert hasattr(pcf.orderers, "MSTOrderer")
+
+
+class TestMSTConformance:
+    """Tests for MST conformance."""
+
+    def test_returns_orderingresult_with_backbone(self):
+        """Returns orderingresult with backbone."""
+        pos, vel, _ = _open_arc()
+        res = pcf.orderers.MSTOrderer(k=8, jump_cap=2.0).order(pos, vel)
+        assert isinstance(res, AbstractResult)
+        assert res.backbone is not None
+        assert res.gamma_range == (-1.0, 1.0)
+        # every tracer ordered (no -1) for a connected graph
+        assert jnp.all(res.indices >= 0)
+        assert set(res.indices.tolist()) == set(range(len(res.indices)))
+        out = res(jnp.linspace(-1.0, 1.0, 11))
+        assert jnp.all(jnp.isfinite(out["x"]))
+
+
+class TestMSTOrderingCorrectness:
+    """Tests for MST ordering correctness."""
+
+    def test_monotone_in_true_arclength(self):
+        """Monotone in true arclength."""
+        pos, vel, t = _open_arc(n=300)
+        res = pcf.orderers.MSTOrderer(k=8, jump_cap=2.0).order(pos, vel)
+        t_ordered = np.asarray(t)[np.asarray(res.ordering)]
[…]
