+++ src/phasecurvefit/_src/orderers/mst.py
@@ -1,0 +1,171 @@
+"""The MST-backbone orderer.
+
+Orders tracers along a 1-D manifold via the longest path (graph diameter) of the
+minimum spanning tree of a kNN graph. Unlike the velocity-following walk, it has
+no progenitor and no forward/backward split: the graph diameter finds the two
+tips itself, giving a clean tip-to-tip ordering. This is the algorithm of choice
+for near-closed-loop / self-overlapping streams where the velocity field reverses
+and a single walk cannot traverse the arc.
+
+The computation is **host-side** (NumPy/SciPy) and deterministic; ``order()`` is
+not jit/vmap-traceable (per the ``AbstractOrderer`` contract).
+"""
+
+__all__: tuple[str, ...] = ("MSTOrderer",)
+
+import warnings
+from typing import Literal
+
+import equinox as eqx
+import jax.numpy as jnp
+import numpy as np
+import plum
+from scipy.sparse import csr_matrix
+from scipy.sparse.csgraph import (
+    connected_components,
+    minimum_spanning_tree,
+    shortest_path,
+)
+from scipy.spatial import cKDTree
+
+from .base import AbstractOrderer
+from .result import OrderingResult
+from phasecurvefit._src.algorithm import StateMetadata
+from phasecurvefit._src.custom_types import VectorComponents
+
+OnDisconnected = Literal["raise", "warn", "largest"]
+
+
+def _backbone_on_component(
+    P: np.ndarray, tree: csr_matrix, nodes: np.ndarray
+) -> tuple[np.ndarray, np.ndarray]:
+    """Backbone/order within one connected component.
+
+    Returns ``(order_idx, Cb)``: original point indices in arc order and the
+    tip-to-tip backbone polyline coordinates.
+    """
+    sub = tree[nodes][:, nodes]
+    # graph diameter via double shortest-path: farthest node a, then farthest b
+    d0 = shortest_path(sub, method="D", indices=0)
+    a = int(np.nanargmax(np.where(np.isinf(d0), -1.0, d0)))
+    da, pred = shortest_path(sub, method="D", indices=a, return_predecessors=True)
+    b = int(np.nanargmax(np.where(np.isinf(da), -1.0, da)))
+    # walk predecessors b -> a to recover the backbone path
+    bb: list[int] = []
+    j = b
+    while j != a and j >= 0:
+        bb.append(j)
+        j = int(pred[j])
+    bb.append(a)
+    bb_local = np.asarray(bb[::-1])  # tip a -> tip b, local indices into ``nodes``
+
+    backbone_nodes = nodes[bb_local]
+    Cb = P[backbone_nodes]  # backbone polyline coordinates
+    seg = np.linalg.norm(np.diff(Cb, axis=0), axis=1)
+    s_bb = np.concatenate([[0.0], np.cumsum(seg)])
+    # project every component point onto the backbone -> along-track arc length
+    _, near = cKDTree(Cb).query(P[nodes])
+    order_local = np.argsort(s_bb[near], kind="stable")
+    return nodes[order_local], Cb
+
+
+def _mst_backbone(
+    P: np.ndarray,
+    *,
+    k: int,
+    jump_cap: float,
+    on_disconnected: OnDisconnected,
+) -> tuple[np.ndarray, np.ndarray]:
+    """Order points along the MST longest-path backbone.
+
+    Returns ``(order_idx, backbone)``: the arc-length ordering (original point
+    indices) and the tip-to-tip backbone polyline coordinates.
+    """
+    n = len(P)
+    if n < 2:
+        return np.arange(n), P.copy()
+
[…]
