### 2026-07-02 0e2d5ac
✨ feat: velocity-aware MST mechanisms
Add three opt-in velocity mechanisms to `MSTOrderer` (pure-spatial stays the
default), all reusing the phase-space velocity alignment cos(v_i, v_j):
- `velocity_weight`: phase-space edge weights that penalise anti-parallel arms;
- `sever_cos_threshold`: drop anti-parallel edges (cuts a reversal seam / a
  self-intersection crossing);
- `orient_by_velocity`: orient gamma along the mean velocity.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
