+++ src/phasecurvefit/_src/orderers/mst.py
@@ -7,6 +7,18 @@ tips itself, giving a clean tip-to-tip ordering. This is the algorithm of choice
 for near-closed-loop / self-overlapping streams where the velocity field reverses
 and a single walk cannot traverse the arc.
 
+Velocity information is **opt-in** (pure-spatial is the default) via three
+mechanisms, all reusing the phase-space notion of velocity alignment
+``cos(v_i, v_j)`` (cf. ``AlignedMomentumDistanceMetric``):
+
+1. *phase-space edge weights* (``velocity_weight``): edge weight
+   ``||dq|| + velocity_weight * (1 - cos(v_i, v_j))`` — anti-parallel arms cost
+   more, so the MST avoids bridging spatially-close arms that move oppositely;
+2. *velocity-aware severing* (``sever_cos_threshold``): drop edges with
+   ``cos(v_i, v_j) < threshold`` — cuts the reversal seam of a near-closed loop;
+3. *tip orientation* (``orient_by_velocity``): flip the ordering so ``gamma``
+   increases along the mean velocity.
+
 The computation is **host-side** (NumPy/SciPy) and deterministic; ``order()`` is
 not jit/vmap-traceable (per the ``AbstractOrderer`` contract).
 """
@@ -34,15 +46,25 @@ from phasecurvefit._src.algorithm import StateMetadata
 from phasecurvefit._src.custom_types import VectorComponents
 
 OnDisconnected = Literal["raise", "warn", "largest"]
+_TINY = 1e-12
+
+
+def _edge_cosine(V: np.ndarray, rows: np.ndarray, cols: np.ndarray) -> np.ndarray:
+    """Cosine similarity of velocities across each candidate edge (i, j)."""
+    vi, vj = V[rows], V[cols]
+    num = np.sum(vi * vj, axis=1)
+    den = np.linalg.norm(vi, axis=1) * np.linalg.norm(vj, axis=1)
+    return np.where(den > _TINY, num / np.maximum(den, _TINY), 0.0)
 
 
 def _backbone_on_component(
     P: np.ndarray, tree: csr_matrix, nodes: np.ndarray
-) -> tuple[np.ndarray, np.ndarray]:
+) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
     """Backbone/order within one connected component.
 
-    Returns ``(order_idx, Cb)``: original point indices in arc order and the
-    tip-to-tip backbone polyline coordinates.
+    Returns ``(order_idx, backbone_nodes, Cb)``: original point indices in arc
+    order, the original indices of the backbone vertices, and the tip-to-tip
+    backbone polyline coordinates.
     """
     sub = tree[nodes][:, nodes]
     # graph diameter via double shortest-path: farthest node a, then farthest b
@@ -66,14 +88,18 @@ def _backbone_on_component(
     # project every component point onto the backbone -> along-track arc length
     _, near = cKDTree(Cb).query(P[nodes])
     order_local = np.argsort(s_bb[near], kind="stable")
-    return nodes[order_local], Cb
+    return nodes[order_local], backbone_nodes, Cb
 
 
 def _mst_backbone(
     P: np.ndarray,
+    V: np.ndarray,
     *,
     k: int,
     jump_cap: float,
+    velocity_weight: float,
+    sever_cos_threshold: float | None,
+    orient_by_velocity: bool,
     on_disconnected: OnDisconnected,
 ) -> tuple[np.ndarray, np.ndarray]:
     """Order points along the MST longest-path backbone.
@@ -94,8 +120,19 @@ def _mst_backbone(
     cols = nn_i[:, 1:].ravel()
     d_edges = nn_d[:, 1:].ravel()  # spatial edge length
 
+    # velocity alignment (only computed when a mechanism needs it)
[…]
