### 2026-07-02 7d57e0f
✨ feat: unxt support for orderers
Mirror `walk_local_flow`'s Quantity-in/Quantity-out UX for the orderers via
plum method-dispatch overloads in the unxt interop. Because MST is host-side,
unit handling is a simple strip-into-usys / reattach-out (no scan_p dispatches):
positions/velocities keep their input units and the backbone is returned in the
position units. `LocalFlowOrderer` delegates to the `walk_local_flow` Quantity
path.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
