### 2026-07-02 04168e5
📝 docs: MST tutorials, orderers guide, and autoencoder integration
Broaden `train_autoencoder`'s dispatch from `WalkLocalFlowResult` to the
unified `OrderingResult` so any orderer's result feeds the AE unchanged, with
an integration test (encoder gamma monotone in MST arc length). Add the
`docs/guides/orderers.md` guide and two tutorials: `epitrochoid_mst`
(spatial vs velocity-aware MST on a self-intersecting curve) and `stream_mst`
(MST on a mock stream, into the autoencoder).

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
