+++ docs/guides/orderers.md
@@ -1,0 +1,97 @@
+# Orderers
+
+An **orderer** turns phase-space tracers `(positions, velocities)` into an
+ordered result that the autoencoder consumes unchanged. All orderers share one
+interface — {class}`~phasecurvefit.orderers.AbstractOrderer` — and return a
+unified {class}`~phasecurvefit.orderers.OrderingResult`, so they are
+interchangeable at call sites:
+
+<!-- skip: next -->
+```python
+import phasecurvefit as pcf
+
+orderer = pcf.orderers.MSTOrderer(k=10, jump_cap=3.0)
+result = orderer.order(qs, ps)  # or: pcf.order(qs, ps, orderer)
+model = pcf.nn.PathAutoencoder.make(normalizer, gamma_range=result.gamma_range, key=key)
+ae, *_ = pcf.nn.train_autoencoder(model, result, config=cfg, key=key)
+```
+
+## Choosing an orderer
+
+| Orderer | Best for | Mechanism |
+|---|---|---|
+| {class}`~phasecurvefit.orderers.LocalFlowOrderer` | open streams; multi-petal / self-intersecting curves where a coherent velocity field can be *followed* | velocity-following greedy walk (wraps {func}`~phasecurvefit.walk_local_flow`) |
+| {class}`~phasecurvefit.orderers.MSTOrderer` | **near-closed loops** and self-overlapping streams where the velocity field *reverses* and a single walk cannot traverse the arc | kNN graph → minimum spanning tree → longest-path (diameter) backbone → arc-length ordering |
+
+The two are complementary. The walk needs a start point and follows the flow; it
+covers only one arm when the velocity reverses at a progenitor. The MST needs no
+progenitor — the graph diameter finds the two tips itself — and orders tip-to-tip
+with bounded per-step jumps, which is exactly what a near-closed loop needs.
+
+## MSTOrderer
+
+The MST is **host-side** (NumPy/SciPy): `order()` is a one-shot preprocessing
+step, not a jit/vmap-traceable function. Pure-spatial is the default:
+
+```python
+import jax.numpy as jnp
+
+import phasecurvefit as pcf
+
+pos = {"x": jnp.linspace(0.0, 5.0, 20), "y": jnp.zeros(20)}
+vel = {"x": jnp.ones(20), "y": jnp.zeros(20)}
+res = pcf.orderers.MSTOrderer(k=5, jump_cap=1.0).order(pos, vel)
+assert res.gamma_range == (-1.0, 1.0)
+assert int(res.n_visited) == 20
+```
+
+`jump_cap` severs edges longer than its value before building the MST; it should
+exceed the typical inter-tracer spacing but stay below the loop-opening /
+arm-separation scale. If the kNN graph is disconnected (e.g. `jump_cap` too
+small), `on_disconnected` controls the response: `"raise"` (default), `"warn"`
+(order the largest component, leave the rest unvisited), or `"largest"` (same,
+silently).
+
+### Velocity is opt-in
+
+Three mechanisms bring velocity into the MST (all off by default), each reusing
+the phase-space notion of velocity alignment `cos(v_i, v_j)`:
+
+- **`velocity_weight`** — edge weights become
+  `||dq|| + velocity_weight * (1 - cos(v_i, v_j))`, so spatially-close arms that
+  move oppositely are not bridged. Set it on the scale of the inter-tracer
+  spacing.
+- **`sever_cos_threshold`** — drop edges with `cos(v_i, v_j)` below the
+  threshold, cutting the reversal seam of a near-closed loop (or cross-branch
+  edges at a self-intersection).
+- **`orient_by_velocity`** — flip the ordering so `gamma` increases along the
+  mean velocity, giving a deterministic, physically-meaningful direction.
+
[…]
