### 2026-07-02 dd95b38
🔧 build: make scipy an optional 'mst' extra, imported lazily
scipy is only needed by MSTOrderer, so move it out of the core dependencies
into an optional `phasecurvefit[mst]` extra (mirroring how `jaxkd` backs the
`kdtree` extra). `mst.py` now imports scipy lazily with a helpful error, so
`import phasecurvefit` works without it and only `MSTOrderer.order()` requires
it. Gate the MST tests on scipy (importorskip) and add scipy to the test/docs
dependency groups so CI still exercises them.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
