### 2026-07-02 3f00717
📝 docs: README orderer examples and a v0.1 migration guide
Add a README "Orderers" section (LocalFlowOrderer vs MSTOrderer, the `pcf.order`
façade) and note the `mst` extra. Add a `docs/migration.md` page covering the
new orderer API, the unified `OrderingResult` (with `WalkLocalFlowResult` now a
subclass), the broadened `train_autoencoder` dispatch, the optional scipy `mst`
extra, and a backward-compatibility summary.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
