### 2026-07-02 bdfc0b5
🐛 fix: address PR review (validate ranges/policies, fix docstring)
- `OrderingResult.__check_init__` rejects a degenerate `gamma_range` (zero
  width) at construction instead of producing inf/NaN in `__call__`.
- `MSTOrderer.__check_init__` rejects unknown `on_disconnected` policies at
  construction instead of silently behaving like "largest".
- Correct the `AbstractOrderer` contract docstring: orderers return `jnp`
  arrays, or unit-aware `unxt` Quantities under the `unxt` interop.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
