### 2026-07-02 ded9e74
🐛 fix: address PR review (docstring accuracy, empty-backbone guard)
- Correct the `OrderingResult.__call__` docstring: `gamma` outside `gamma_range`
  raises (via `eqx.error_if`); it is not clipped.
- Raise a clear `ValueError` when interpolating along an empty backbone
  (the degenerate n=0 case) instead of crashing with an out-of-bounds index.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
