### 2026-07-04 1fe6011
✨ feat: swappable trainable decoder + FourierTrackNet
Add an `AbstractTrackNet` interface for the trainable gamma->position decoder
and a `FourierTrackNet` variant that feeds Fourier features
[gamma, sin(pi k gamma), cos(pi k gamma)] to the MLP, giving it the
high-frequency capacity to represent sharp / self-intersecting tracks a plain
`TrackNet` over-smooths. `TrackNet` is re-parented onto the new base; both are
interchangeable via a new `PathAutoencoder.make(decoder=...)` override
(validates out_size; defaults to `TrackNet` for backward compatibility).

Update the epitrochoid MST tutorial to
[…]
