+++ src/phasecurvefit/_src/nn/track_net.py
@@ -1,7 +1,13 @@
 r"""Interpolation Network for interpolating skipped tracers."""
 
-__all__: tuple[str, ...] = ("TrackNet", "decoder_loss")
+__all__: tuple[str, ...] = (
+    "AbstractTrackNet",
+    "TrackNet",
+    "FourierTrackNet",
+    "decoder_loss",
+)
 
+import abc
 import functools as ft
 from dataclasses import KW_ONLY, dataclass
 from typing import TypeAlias, cast
@@ -19,7 +25,37 @@ from .utils import masked_mean, shuffle_and_batch
 from phasecurvefit._src.custom_types import FSz0, RSz0, RSzN
 
 
-class TrackNet(eqx.Module):
+class AbstractTrackNet(eqx.Module):
+    r"""Interface for a trainable decoder mapping $\gamma \to$ position.
+
+    A track net is the second half of the autoencoder: it reconstructs the
+    spatial track position from the scalar ordering parameter $\gamma$.
+
+    Subclasses must:
+
+    - expose ``out_size`` (the number of spatial dimensions), and
+    - implement ``__call__(gamma, /, key=None) -> position`` of shape
+      ``(out_size,)``, **differentiable in** ``gamma`` (Phase-2 training takes a
+      ``jvp`` of the decoder w.r.t. $\gamma$ for the velocity/tangent loss).
+
+    Concrete variants (all interchangeable via ``PathAutoencoder.make(decoder=...)``):
+
+    - :class:`TrackNet` — a plain MLP on $\gamma$ (default).
+    - :class:`FourierTrackNet` — Fourier features on $\gamma$, for sharp /
+      self-intersecting tracks a plain MLP over-smooths.
+    """
+
+    out_size: eqx.AbstractVar[int]
+
+    @abc.abstractmethod
+    def __call__(
+        self, gamma: RSz0, /, key: PRNGKeyArray | None = None
+    ) -> Float[Array, " {self.out_size}"]:
+        """Reconstruct the position at ordering parameter ``gamma``."""
+        ...
+
+
+class TrackNet(AbstractTrackNet):
     r"""Param-Net (decoder): maps $\gamma \to$ position (x, y, z).
 
     This network reconstructs the stream track position from the ordering
@@ -93,6 +129,75 @@ class TrackNet(eqx.Module):
 
         """
         return self.mlp(gamma, key=key)
+
+
+class FourierTrackNet(AbstractTrackNet):
+    r"""Decoder with Fourier features on $\gamma$ before the MLP.
+
+    Maps $\gamma$ to ``[gamma, sin(pi k gamma), cos(pi k gamma)]`` for
+    ``k = 1 .. n_frequencies`` and feeds those ``1 + 2*n_frequencies`` features
+    to an MLP. The Fourier features give the network the high-frequency capacity
+    to represent sharp or self-intersecting tracks (e.g. multi-petal curves)
+    that a plain :class:`TrackNet` tends to over-smooth. ``sin/cos(pi k gamma)``
+    complete one period over a unit-width $\gamma$-range, so ``n_frequencies``
+    controls how sharp a curve the decoder can render.
+
+    Parameters
+    ----------
+    out_size : int
+        Number of spatial dimensions.
+    n_frequencies : int
+        Number of Fourier modes ``k``. Higher captures sharper structure.
+    width_size, depth : int
+        MLP hidden-layer size and number of hidden layers.
+    key : PRNGKeyArray
+        JAX random key for initialization.
+
+    """
+
+    mlp: eqx.nn.MLP
+
+    out_size: int = eqx.field(static=True)
+    n_frequencies: int = eqx.field(static=True)
+    width_size: int = eqx.field(static=True)
[…]
