### 2026-07-04 e15fec4
🐛 fix: train the AE decoder in the ordering's gamma, not the noisy encoder gamma
Phase-2 built the decoder's running-mean target by denoising over the encoder's
gamma. That gamma is globally monotone but jittery at the running-mean window
scale, so the window averaged spatially-scattered points and blurred the target
to ~8x the tracer spacing off the track. The decoder then faithfully learned the
blur (loss ~2e-4, yet the mean path sat far from the data) -- so the "bad
decoder" was really a bad target.

Denoise and train the decoder in the ORDERING's gamma (exact rank along the
walk) instead; 
[…]
