### 2026-07-04 de2dc89
✅ test: regression test for MSTOrderer positions/velocities key mismatch
Locks in the clear `ValueError` (added in "Apply suggestions from code review")
when `positions` and `velocities` have mismatched component keys, instead of a
cryptic `KeyError`.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
