### 2026-07-05 0a9b4e3
📝 docs: document LocalFlowOrderer in the orderers guide + MST cross-refs
- orderers guide: add a full `LocalFlowOrderer` section (the velocity-following
  walk) parallel to the MST section — what it does, its hyperparameters
  (metric_scale/config/start_idx/direction/max_dist), and when to prefer it vs
  the MST. Previously only MST had a dedicated section.
- Cross-reference the MST orderer where relevant in the other guides: algorithm
  (walk is one orderer; MST for reversing-velocity loops), metrics (MST reuses
  the cos-theta alignment idea), jax-integration (MST is host-side, not
  traceab
[…]
