### 2026-07-05 2a858e8
📝 docs: use the pcf.order() API in the algorithm guide examples
Rewrite the walk examples in the algorithm guide to run through the newer
`pcf.order(pos, vel, LocalFlowOrderer(...))` entry point (still the local-flow
walk), and note that `walk_local_flow` remains available and equivalent.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
