### 2026-07-05 0fadbeb
📝 docs: use the pcf.order() API in the jax, metrics, and quickstart guides
Rewrite every walk example in these guides to run through the newer
`pcf.order(pos, vel, LocalFlowOrderer(...))` entry point instead of the
`walk_local_flow` function (including the vmap `static_kw={"orderer": ...}`
form, the grad example, and the unxt `metadata=StateMetadata(usys=...)` path).
Note that `walk_local_flow` remains available and equivalent.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
