+++ docs/guides/metrics.md
@@ -1,8 +1,8 @@
 # Distance Metrics Guide
 
-The `walk_local_flow` algorithm uses distance metrics to determine how to select
-the next point in a phase-space trajectory. This guide explains the metric
-system and shows how to use and create custom metrics.
+The local-flow walk uses distance metrics to determine how to select the next
+point in a phase-space trajectory. This guide explains the metric system and
+shows how to use and create custom metrics.
 
 ## Overview
 
@@ -31,7 +31,11 @@ pos = {"x": jnp.array([0.0, 1.0, 2.0]), "y": jnp.array([0.0, 0.5, 1.0])}
 vel = {"x": jnp.array([1.0, 1.0, 1.0]), "y": jnp.array([0.5, 0.5, 0.5])}
 
 config = pcf.WalkConfig(metric=FullPhaseSpaceDistanceMetric())
-result = pcf.walk_local_flow(pos, vel, config=config, start_idx=0, metric_scale=1.0)
+result = pcf.order(
+    pos,
+    vel,
+    pcf.orderers.LocalFlowOrderer(config=config, start_idx=0, metric_scale=1.0),
+)
 ```
 
 ## Built-in Metrics
@@ -60,7 +64,11 @@ from phasecurvefit.metrics import SpatialDistanceMetric
 
 # Pure nearest-neighbor search in position space
 config = pcf.WalkConfig(metric=SpatialDistanceMetric())
-result = pcf.walk_local_flow(pos, vel, config=config, start_idx=0, metric_scale=0.0)
+result = pcf.order(
+    pos,
+    vel,
+    pcf.orderers.LocalFlowOrderer(config=config, start_idx=0, metric_scale=0.0),
+)
 ```
 
 ### AlignedMomentumDistanceMetric
@@ -94,7 +102,11 @@ vel = {"x": jnp.array([1.0, 1.0, 1.0]), "y": jnp.array([0.5, 0.5, 0.5])}
 
 # Aligned momentum metric
 config = pcf.WalkConfig(metric=AlignedMomentumDistanceMetric())
-result = pcf.walk_local_flow(pos, vel, config=config, start_idx=0, metric_scale=1.0)
+result = pcf.order(
+    pos,
+    vel,
+    pcf.orderers.LocalFlowOrderer(config=config, start_idx=0, metric_scale=1.0),
+)
 ```
 
 ### FullPhaseSpaceDistanceMetric
@@ -132,7 +144,11 @@ from phasecurvefit.metrics import FullPhaseSpaceDistanceMetric
 # Full 6D phase-space distance (this is the default)
 # metric_scale represents a time scale (e.g., if pos ~ kpc, vel ~ kpc/Myr, metric_scale ~ Myr)
 config = pcf.WalkConfig(metric=FullPhaseSpaceDistanceMetric())
-result = pcf.walk_local_flow(pos, vel, config=config, start_idx=0, metric_scale=1.0)
+result = pcf.order(
+    pos,
+    vel,
+    pcf.orderers.LocalFlowOrderer(config=config, start_idx=0, metric_scale=1.0),
+)
 ```
 
 **Comparison with momentum metric:**
@@ -212,7 +228,11 @@ pos = {"x": jnp.array([0.0, 1.0, 2.0]), "y": jnp.array([0.0, 0.5, 1.0])}
 vel = {"x": jnp.array([1.0, 1.0, 1.0]), "y": jnp.array([0.5, 0.5, 0.5])}
 
 config = pcf.WalkConfig(metric=Full6DMetric())
-result = pcf.walk_local_flow(pos, vel, config=config, start_idx=0, metric_scale=1.0)
+result = pcf.order(
+    pos,
+    vel,
[…]
