### 2026-07-05 d42f0f4
📝 docs: drop walk_local_flow mentions from the guides in favor of pcf.order
Remove the remaining `walk_local_flow` references (prose, xrefs, and the last
call in the autoencoder guide) from the updated guides; they now present the
local-flow walk solely via `pcf.order` + `LocalFlowOrderer`.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
