+++ docs/guides/algorithm.md
@@ -6,10 +6,9 @@ the phase flow walking algorithm.
 ```{note}
 The walk described here is one of the pluggable **orderers**. The examples below
 run it via `pcf.order` with a
-{class}`~phasecurvefit.orderers.LocalFlowOrderer`; the historical
-{func}`~phasecurvefit.walk_local_flow` remains available and equivalent. The walk
-*follows* the velocity field, so it covers only one arm of a **near-closed loop**
-whose velocity reverses at a progenitor. For that case the
+{class}`~phasecurvefit.orderers.LocalFlowOrderer`. The walk *follows* the
+velocity field, so it covers only one arm of a **near-closed loop** whose
+velocity reverses at a progenitor. For that case the
 {class}`~phasecurvefit.orderers.MSTOrderer` backbone orders the loop tip-to-tip
 without a start point — see the [Orderers guide](orderers.md).
 ```
