### 2026-07-05 fd26ecd
Update docs/migration.md
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
