### 2026-07-05 fac679f
Update docs/migration.md
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
