### 2026-07-06 c5d4ecd
📚 docs: add example sections for OrderingResult, MSTOrderer, and FourierTrackNet
Signed-off-by: nstarman <nstarman@users.noreply.github.com>
