### 2026-07-06 fa34cfa
✨ feat: pluggable orderer abstraction with a velocity-aware MST backbone (#35)
✨ feat: pluggable orderers, MST ordering, Fourier decoder, unified API

- Introduce pluggable `AbstractOrderer` interface, `pcf.order()` façade, and
  unified `OrderingResult` (`WalkLocalFlowResult` retained as a subclass for
  compatibility).
- Add `MSTOrderer` with spatial and optional velocity-aware ordering, plus
  `unxt` support and optional SciPy `mst` extra.
- Generalize the autoencoder to accept any `OrderingResult`; add swappable
  `AbstractTrackNet`/`FourierTrackNet` decoder support.
- Fix decoder training
[…]
