+++ src/phasecurvefit/_src/orderers/result.py
@@ -1,0 +1,233 @@
+"""Unified result type for ordering algorithms.
+
+``OrderingResult`` is the single concrete result returned by orderers. It holds
+the four fields the autoencoder consumes (``positions``, ``velocities``,
+``indices``, ``gamma_range``) plus an optional ``backbone`` polyline, and a
+single backbone-aware ``__call__``:
+
+- with a ``backbone`` (e.g. from ``MSTOrderer``): interpolate along the smooth
+  tip-to-tip polyline (a clean centerline);
+- without one (e.g. from the local-flow walk): interpolate along the ordered
+  visited observations, reproducing the legacy walk behavior.
+
+``WalkLocalFlowResult`` is a thin subclass of this class (see ``algorithm``),
+preserving its public name while sharing all logic.
+"""
+
+__all__: tuple[str, ...] = ("OrderingResult",)
+
+from dataclasses import KW_ONLY
+
+import equinox as eqx
+import jax.numpy as jnp
+import jax.tree as jt
+from jaxtyping import Array, Bool, Int, PRNGKeyArray
+
+from zeroth import zeroth
+
+from phasecurvefit._src.abstract_result import AbstractResult
+from phasecurvefit._src.custom_types import BSzN, ISz0, ISzN, VectorComponents
+
+
+class OrderingResult(AbstractResult):
+    r"""Unified result of an ordering algorithm.
+
+    Attributes
+    ----------
+    positions : dict[str, Array]
+        Original (not reordered) position components, 1D arrays of shape
+        ``(n_obs,)``.
+    velocities : dict[str, Array]
+        Original velocity components, same shape as ``positions``.
+    indices : Int[Array, " n_obs"]
+        Ordered indices of visited observations; unvisited slots are ``-1``.
+    gamma_range : tuple[float, float]
+        Static valid range of the ordering parameter for ``__call__``.
+    backbone : dict[str, Array] | None
+        Optional ordered polyline (tip-to-tip) that ``__call__`` interpolates
+        along. ``None`` for walk-style results, which interpolate along the
+        ordered visited observations instead.
+
+    Examples
+    --------
+    Create an ``OrderingResult`` from ordering observations:
+
+    >>> import jax.numpy as jnp
+    >>> import phasecurvefit as pcf
+
+    >>> # Define phase-space data
+    >>> positions = {
+    ...     "x": jnp.array([0.0, 1.0, 2.0, 3.0]),
+    ...     "y": jnp.array([0.0, 0.5, 1.0, 1.5]),
+    ... }
+    >>> velocities = {"x": jnp.ones(4), "y": jnp.full(4, 0.5)}
+    >>> indices = jnp.array([0, 1, 2, 3])  # All observations visited
+
+    >>> # Create a result object
+    >>> result = pcf.orderers.OrderingResult(
+    ...     positions=positions,
+    ...     velocities=velocities,
+    ...     indices=indices,
+    ...     gamma_range=(0.0, 1.0),
+    ... )
+
+    Access ordering properties:
+
+    >>> result.ordering  # Ordered indices
+    Array([0, 1, 2, 3], dtype=int32)
+    >>> result.n_visited  # Number of observations included
+    Array(4, dtype=int32)
+    >>> result.n_skipped  # Number of observations skipped
+    Array(0, dtype=int32)
+    >>> result.all_visited  # Were all observations included?
+    Array(True, dtype=bool)
+
+    Get ordered data:
+
+    >>> ordered_pos, ordered_vel = result.ordered
+    >>> ordered_pos["x"]
+    Array([0., 1., 2., 3.], dtype=float32)
+
+    Interpolate positions along the ordering using ``gamma``:
+
+    >>> # gamma=0 returns the first position, gamma=1 the last
+    >>> result(jnp.array(0.0))
[…]
