### 2026-07-06 7a0b2dc
📝 Fix docstring indentation in FullPhaseSpaceDistanceMetric
- Convert bullet list to inline text in docstring
- Fixes ReadTheDocs build error: 'Unexpected indentation'

Signed-off-by: nstarman <nstarman@users.noreply.github.com>
