### 2026-07-06 6e5d8b4
Clarify metric description in metrics.py
Updated the metric description to clarify the roles of distance in position and velocity space.

Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
