### 2026-07-06 1c0e97f
🔧 Clarify scipy as core dependency in docs and remove mst extra
- scipy>=1.11 is now a required core dependency (not optional)
- Remove mst extra from [project.optional-dependencies]
- Update 'all' extra to exclude mst (which no longer exists)
- Remove scipy from docs and test dependency-groups (already in core)
- Update README to reflect scipy as required, not optional
- Installation instructions now reflect actual dependency structure
