### 2026-07-06 13fabe4
📦 Make scipy a required dependency (#38)
* 📦 Make scipy a required dependency and refactor imports

- Move scipy>=1.11 from optional[mst] to core dependencies
- Remove lazy scipy loading from mst.py
- Use direct scipy imports at module level
- Update optional-dependencies to remove mst extra
- Remove pytest.importorskip('scipy') from tests
- Update README to reflect scipy as required

* 📝 Fix docstring indentation in FullPhaseSpaceDistanceMetric

- Convert bullet list to inline text in docstring
- Fixes ReadTheDocs build error: 'Unexpected indentation'
- Set nbsphinx_execute = 'never' when REA
[…]
