### 2026-07-13 5c9e2c9
👽️ external: refactor ML training using jaxmore.nn>=v0.4.0
Signed-off-by: nstarman <nstarman@users.noreply.github.com>
