### 2026-07-13 fe4103d
👽️ external: refactor ML training using jaxmore.nn>=v0.4.0
Signed-off-by: nstarman <nstarman@users.noreply.github.com>
