+++ src/phasecurvefit/_src/nn/autoencoder.py
@@ -1,20 +1,27 @@
 """Autoencoder."""
 
-__all__: tuple[str, ...] = ("PathAutoencoder", "train_autoencoder", "TrainingConfig")
+__all__: tuple[str, ...] = (
+    "PathAutoencoder",
+    "PathAutoencoderTrainer",
+    "train_autoencoder",
+    "TrainingConfig",
+)
 
+import functools as ft
 from collections.abc import Mapping
 from dataclasses import KW_ONLY, dataclass
-from typing import ClassVar, TypeAlias, cast
+from typing import Any, ClassVar, TypeAlias, cast
 
 import equinox as eqx
 import jax
 import jax.numpy as jnp
 import jax.random as jr
 import jax.tree as jtu
-import jax_tqdm
 import optax
 import plum
 from jaxtyping import Array, Bool, Float, Int, PRNGKeyArray, PyTree
+
+from jaxmore.nn import masked_mean
 
 from .abstractautoencoder import AbstractAutoencoder
 from .externaldecoder import RunningMeanDecoder
@@ -33,7 +40,7 @@ from .track_net import (
     decoder_loss,
     train_track_net,
 )
-from .utils import shuffle_and_batch
+from .trainer import AbstractEqxScanTrainer, EqxTrainCarry
 from phasecurvefit._src.custom_types import FLikeSz0, FSz0, FSzN
 from phasecurvefit._src.orderers.result import OrderingResult
 
@@ -485,13 +492,135 @@ def make_step(
 # ===================================================================
 
 
-BatchScanCarry: TypeAlias = tuple[eqx.Module, optax.OptState, PRNGKeyArray]  # noqa: UP040
-BatchScanInputs: TypeAlias = tuple[  # noqa: UP040
-    Bool[Array, " B"],  # batch_mask: True where batch has data, False where padded
-    Float[Array, " B TwoF"],  # batch_ws : ordered positions
-    Float[Array, " B"],  # batch_weights:
-    FSz0,  # lambda_p: scalar array
-]
+def _pad_to_multiple(
+    mask: Bool[Array, " N"],
+    *args: Float[Array, "N ..."],
+    batch_size: int,
+    pad_value: float,
+) -> tuple[Bool[Array, " M"], tuple[Float[Array, "M ..."], ...]]:
+    """Pad the dataset up to a whole number of batches, with `pad_value`.
+
+    Why this exists
+    ---------------
+    `jaxmore.nn.AbstractScanNNTrainer.run` calls `shuffle_and_batch` with the
+    default ``pad_value=0``; it does not (yet) expose the parameter. For this
+    loss that is *not* a neutral choice: `compute_decoder_loss` penalises
+    ``~mask`` rows as membership false-positives, and padding rows are ``~mask``.
+    Padding with zeros would therefore plant fabricated "non-member" points at
+    the origin of the normalised phase space -- i.e. right where the stream is --
+    and actively train the encoder to reject its own centre.
+
+    By padding here to an exact multiple of `batch_size`, `shuffle_and_batch`
+    finds nothing left to pad and `pad_value` is never consulted. The rows we add
+    carry the original ``pad_value=1`` and ``mask=False``, exactly as before.
+
+    Delete this once `run()` accepts `pad_value`.
+
+    """
+    n = len(mask)
+    n_batches = -(-n // batch_size)  # ceiling division
+    pad_amount = n_batches * batch_size - n
+    if pad_amount == 0:
+        return mask, args
+
+    padded_mask = jnp.pad(mask, (0, pad_amount), constant_values=False)
+    padded_args = tuple(
+        jnp.pad(
+            arr,
+            [(0, pad_amount)] + [(0, 0)] * (arr.ndim - 1),
+            constant_values=pad_value,
+        )
+        for arr in args
+    )
[…]
