+++ src/phasecurvefit/_src/nn/track_net.py
@@ -4,24 +4,26 @@ __all__: tuple[str, ...] = (
     "AbstractTrackNet",
     "TrackNet",
     "FourierTrackNet",
+    "TrackNetTrainer",
     "decoder_loss",
 )
 
 import abc
 import functools as ft
 from dataclasses import KW_ONLY, dataclass
-from typing import TypeAlias, cast
+from typing import Any, cast
 
 import equinox as eqx
 import jax
 import jax.numpy as jnp
 import jax.random as jr
-import jax_tqdm
 import optax
 from jaxtyping import Array, Bool, Float, PRNGKeyArray, Real
 
+from jaxmore.nn import masked_mean
+
 from .order_net import default_optimizer
-from .utils import masked_mean, shuffle_and_batch
+from .trainer import AbstractEqxScanTrainer, EqxTrainCarry
 from phasecurvefit._src.custom_types import FSz0, RSz0, RSzN
 
 
@@ -416,12 +418,62 @@ class TrackTrainingConfig:
     """Whether to show an epoch progress bar via tqdm."""
 
 
-BatchScanCarry: TypeAlias = tuple[eqx.Module, optax.OptState, PRNGKeyArray]  # noqa: UP040
-BatchScanInputs: TypeAlias = tuple[  # noqa: UP040
-    Bool[Array, " B"],  # mask
-    Float[Array, " B"],  # gamma
-    Real[Array, " B D"],  # qs_mean
-]
+def _track_step(
+    carry: EqxTrainCarry,
+    batch_inputs: tuple[Bool[Array, " B"], tuple[Array, ...]],
+    *,
+    optimizer: optax.GradientTransformation,
+    filter_spec: Any,
+) -> tuple[FSz0, EqxTrainCarry]:
+    """Run one batch of TrackNet training.
+
+    `batch_inputs` is ``(mask, (gamma, qs_mean))``.
+
+    `filter_spec` must be the same spec the trainer used to build `opt_state`
+    (see `TrackNetTrainer.init`), so that the step, the carry packing in
+    `AbstractEqxScanTrainer.pack_carry_state`, and the optimizer state all
+    agree on which leaves are trainable.
+    """
+    model, opt_state, key = carry
+    mask, (gamma, qs_mean) = batch_inputs
+
+    model_dynamic, model_static = eqx.partition(model, filter_spec)
+
+    key, subkey = jr.split(key)
+    loss, model_dynamic, opt_state = make_step(
+        model_dynamic,
+        model_static,
+        gamma=gamma,
+        qs_mean=qs_mean,
+        mask=mask,
+        opt_state=opt_state,
+        optimizer=optimizer,
+        key=subkey,
+    )
+
+    model = eqx.combine(model_dynamic, model_static)
+    return loss, (model, opt_state, key)
+
+
+@dataclass(frozen=True)
+class TrackNetTrainer(AbstractEqxScanTrainer):
+    """Scan trainer for `TrackNet`."""
+
+    def init(  # type: ignore[override]
+        self,
+        model: TrackNet,
+        /,
+        *,
+        gamma: Float[Array, " N"],
+        qs_mean: Real[Array, " N D"],
+        mask: Bool[Array, " N"],
+        optimizer: optax.GradientTransformation,
+        key: PRNGKeyArray,
+    ) -> tuple[EqxTrainCarry, tuple[Bool[Array, " N"], tuple[Array, ...]]]:
+        """Build the initial carry and the epoch data."""
+        model_dynamic, _ = eqx.partition(model, self.filter_spec)
+        opt_state = optimizer.init(model_dynamic)
+        return (model, opt_state, key), (mask, (gamma, qs_mean))
 
 
[…]
