+++ pyproject.toml
@@ -20,7 +20,7 @@ classifiers = [
 dependencies = [
   "equinox>=0.13.5",
   "jax>=0.7.2",
-  "jaxmore>=0.2.0",
+  "jaxmore>=0.4.0",
   "jax-tqdm>=0.4.0",
   "jaxtyping>=0.3.5",
   "optax>=0.2.4",
