### 2026-07-13 e5ff405
✨ feat: membership prob of Hogg, Bovy & Lang (2010)
Signed-off-by: nstarman <nstarman@users.noreply.github.com>
