+++ tests/test_autoencoder.py
@@ -42,7 +42,7 @@ class TestOrderingNet:
 
         # Batch of 10 points in 2D phase-space (4 features: x, y, vx, vy)
         key, subkey = jr.split(key)
-        w = jax.random.normal(subkey, (10, 4))
+        w = jr.normal(subkey, (10, 4))
         gamma, prob = jax.vmap(net)(w)
 
         assert gamma.shape == (10,)
@@ -54,7 +54,7 @@ class TestOrderingNet:
         net = pcf.nn.OrderingNet(in_size=4, key=subkey)
 
         key, subkey = jr.split(key)
-        w = jax.random.normal(subkey, (100, 4))
+        w = jr.normal(subkey, (100, 4))
         gamma, _ = jax.vmap(net)(w)
 
         assert jnp.all(gamma >= 0.0)
@@ -66,7 +66,7 @@ class TestOrderingNet:
         net = pcf.nn.OrderingNet(in_size=4, key=subkey)
 
         key, subkey = jr.split(key)
-        w = jax.random.normal(subkey, (100, 4))
+        w = jr.normal(subkey, (100, 4))
         _, prob = jax.vmap(net)(w)
 
         assert jnp.all(prob >= 0.0)
@@ -78,7 +78,7 @@ class TestOrderingNet:
         net = pcf.nn.OrderingNet(in_size=4, key=subkey)
 
         key, subkey = jr.split(key)
-        w = jax.random.normal(subkey, (4,))
+        w = jr.normal(subkey, (4,))
         gamma, prob = net(w)
 
         assert gamma.shape == ()
@@ -332,7 +332,7 @@ class TestTrainAutoencoder:
         normalizer = pcf.nn.StandardScalerNormalizer(
             simple_wlf_result.positions, simple_wlf_result.velocities
         )
-        key1, key2 = jax.random.split(rng_key)
+        key1, key2 = jr.split(rng_key)
         ae = pcf.nn.PathAutoencoder.make(normalizer, gamma_range=(0.0, 1.0), key=key1)
 
         # Use minimal epochs for fast testing
@@ -352,7 +352,7 @@ class TestTrainAutoencoder:
         normalizer = pcf.nn.StandardScalerNormalizer(
             simple_wlf_result.positions, simple_wlf_result.velocities
         )
-        key1, key2 = jax.random.split(rng_key)
+        key1, key2 = jr.split(rng_key)
         ae = pcf.nn.PathAutoencoder.make(normalizer, gamma_range=(0.0, 1.0), key=key1)
 
         # Use more epochs for phase 1 to see loss reduction
@@ -382,7 +382,7 @@ class TestTrainAutoencoder:
         normalizer = pcf.nn.StandardScalerNormalizer(
             simple_wlf_result.positions, simple_wlf_result.velocities
         )
-        key1, key2 = jax.random.split(rng_key)
+        key1, key2 = jr.split(rng_key)
         ae = pcf.nn.PathAutoencoder.make(normalizer, gamma_range=(0.0, 1.0), key=key1)
 
         config = pcf.nn.TrainingConfig(
@@ -407,11 +407,11 @@ class TestFillOrderingGaps:
     def phasecurvefit_with_gaps(self):
         """Create a simple phase-flow walk result with skipped tracers."""
         n_points = 30
-        key = jax.random.key(42)
+        key = jr.key(42)
 
         # Create a curved stream
         theta = jnp.linspace(0, jnp.pi, n_points)
-        shuffle_idx = jax.random.permutation(key, n_points)
+        shuffle_idx = jr.permutation(key, n_points)
 
         pos = {
             "x": jnp.cos(theta)[shuffle_idx],
@@ -437,7 +437,7 @@ class TestFillOrderingGaps:
         normalizer = pcf.nn.StandardScalerNormalizer(
[…]
