+++ tests/test_membership.py
@@ -1,0 +1,483 @@
+"""Tests for mixture-model membership (outlier rejection).
+
+The behavioural tests build a synthetic arc-shaped stream with known outliers
+and check that the model actually finds them. They are the tests that matter:
+the unit tests below them only pin the pieces in place.
+"""
+
+import equinox as eqx
+import jax
+import jax.numpy as jnp
+import jax.random as jr
+import numpy as np
+import optax
+import pytest
+
+import phasecurvefit as pcf
+from phasecurvefit.nn import (
+    MixtureMembershipConfig,
+    StreamWidthNet,
+    membership_rampup,
+    membership_responsibility,
+    mixture_membership_loss,
+    sigma_ceiling,
+    uniform_background_density,
+)
+
+# ============================================================
+# StreamWidthNet
+
+
+class TestStreamWidthNet:
+    """The gamma-dependent stream width network."""
+
+    def test_initialises_near_sigma_init(self) -> None:
+        """The net is constructed to predict ~sigma_init everywhere."""
+        net = StreamWidthNet(sigma_init=0.25, key=jr.key(0))
+        gammas = jnp.linspace(-1.0, 1.0, 32)
+        sigmas = jax.vmap(net)(gammas)
+
+        # Exactly sigma_init + sigma_min, since the final layer is zeroed.
+        assert jnp.allclose(sigmas, 0.25 + net.sigma_min, atol=1e-6)
+
+    def test_strictly_positive(self) -> None:
+        """Widths must never be <= 0, whatever the weights."""
+        net = StreamWidthNet(sigma_init=0.1, sigma_min=1e-3, key=jr.key(1))
+        # Perturb the weights hard; positivity is structural (exp + floor).
+        net = jax.tree_util.tree_map(lambda x: x - 50.0 if eqx.is_array(x) else x, net)
+        sigmas = jax.vmap(net)(jnp.linspace(-5.0, 5.0, 64))
+
+        assert bool(jnp.all(sigmas >= net.sigma_min))
+        assert bool(jnp.all(jnp.isfinite(sigmas)))
+
+    def test_can_vary_along_track(self) -> None:
+        """After training, sigma(gamma) is genuinely gamma-dependent."""
+        net = StreamWidthNet(sigma_init=0.1, key=jr.key(2))
+        gammas = jnp.linspace(-1.0, 1.0, 64)
+        # Target: width doubles along the track (a fanning stream).
+        target = 0.1 + 0.1 * (gammas + 1) / 2
+
+        def loss(net):
+            return jnp.mean((jax.vmap(net)(gammas) - target) ** 2)
+
+        opt = optax.adam(1e-2)
+        st = opt.init(eqx.filter(net, eqx.is_array))
+        for _ in range(500):
+            g = eqx.filter_grad(loss)(net)
+            u, st = opt.update(g, st, eqx.filter(net, eqx.is_array))
+            net = eqx.apply_updates(net, u)
+
+        got = jax.vmap(net)(gammas)
+        assert jnp.allclose(got, target, atol=0.02)
+        assert float(got[-1]) > 1.5 * float(got[0])  # really does vary
+
+    @pytest.mark.parametrize("bad", [0.0, -1.0])
+    def test_rejects_nonpositive_sigma_init(self, bad: float) -> None:
+        """A non-positive initial width is meaningless."""
+        with pytest.raises(ValueError, match="sigma_init must be positive"):
+            StreamWidthNet(sigma_init=bad, key=jr.key(0))
+
+
+# ============================================================
[…]

[…]
