### 2026-07-13 a383e66
👽️ external: refactor ML training using jaxmore.nn>=v0.4.0 (#42)
Signed-off-by: nstarman <nstarman@users.noreply.github.com>
