+++ src/phasecurvefit/_src/nn/__init__.py
@@ -60,4 +60,4 @@ from .normalize import *
 from .order_net import *
 from .result import *
 from .track_net import *
-from .utils import *
+from .trainer import *
