+++ src/phasecurvefit/_src/nn/autoencoder.py
@@ -1,20 +1,27 @@
 """Autoencoder."""
 
-__all__: tuple[str, ...] = ("PathAutoencoder", "train_autoencoder", "TrainingConfig")
+__all__: tuple[str, ...] = (
+    "PathAutoencoder",
+    "PathAutoencoderTrainer",
+    "train_autoencoder",
+    "TrainingConfig",
+)
 
+import functools as ft
 from collections.abc import Mapping
 from dataclasses import KW_ONLY, dataclass
-from typing import ClassVar, TypeAlias, cast
+from typing import Any, ClassVar, TypeAlias, cast
 
 import equinox as eqx
 import jax
 import jax.numpy as jnp
 import jax.random as jr
 import jax.tree as jtu
-import jax_tqdm
 import optax
 import plum
 from jaxtyping import Array, Bool, Float, Int, PRNGKeyArray, PyTree
+
+from jaxmore.nn import masked_mean
 
 from .abstractautoencoder import AbstractAutoencoder
 from .externaldecoder import RunningMeanDecoder
@@ -33,7 +40,7 @@ from .track_net import (
     decoder_loss,
     train_track_net,
 )
-from .utils import shuffle_and_batch
+from .trainer import AbstractEqxScanTrainer, EqxTrainCarry
 from phasecurvefit._src.custom_types import FLikeSz0, FSz0, FSzN
 from phasecurvefit._src.orderers.result import OrderingResult
 
@@ -410,7 +417,15 @@ def compute_decoder_loss(
     # Modify the mask to avoid training on non-members. This actually can't add
     # members (that would be an 'or' operation) and we penalize false positives
     # and negatives above, so this has a smaller effect.
-    mask = mask & is_member
+    #
+    # It can, however, empty the mask entirely, and that is not hypothetical: the
+    # trainer only guarantees the *batch* mask has a usable sample (it skips
+    # batches that do not), while `is_member` is re-evaluated from the live
+    # encoder at every step. Mid-training the encoder can transiently drop every
+    # star in the batch below the threshold, and `masked_mean` is documented to
+    # return NaN on an empty mask -- which would poison the epoch loss. Hence the
+    # guard on `loss_path` below.
+    member_mask = mask & is_member
 
     # Compute dq/dgamma (Jacobian of decoder output w.r.t. gamma) elementwise
     # jax.jacobian gives us the derivative of decoder output w.r.t. its input
@@ -427,15 +442,26 @@ def compute_decoder_loss(
     t_norm = jnp.linalg.norm(dq_dgamma, axis=1, keepdims=True)
     t_hat = jnp.where(t_norm > 0, dq_dgamma / t_norm, jnp.zeros_like(dq_dgamma))
 
-    loss_path = decoder_loss(
-        qs_meas=ws[:, :D],
-        weights=weights,
-        qs_pred=q_predict,
-        t_hat=t_hat,
-        p_hat=ps_hat,
-        mask=mask,
-        lambda_q=lambda_q,
-        lambda_p=lambda_p,
+    def _path_loss() -> FSz0:
+        return decoder_loss(
+            qs_meas=ws[:, :D],
+            weights=weights,
+            qs_pred=q_predict,
+            t_hat=t_hat,
+            p_hat=ps_hat,
+            mask=member_mask,
+            lambda_q=lambda_q,
+            lambda_p=lambda_p,
+        )
+
+    # When `member_mask` is empty the model currently claims no member in this
+    # batch, so there is no track to fit: the path loss is zero and only
+    # `loss_mismember` supplies gradient, pushing the probabilities back up (as
+    # it in fact does -- the collapse is transient). `lax.cond` rather than
[…]

[…]
