### 2026-07-13 3ab4498
🐛 fix: green up CI doctests for membership guide/module
Sybil executes the code blocks in docs and docstrings. Three illustrative
`{code-block} python` snippets in the outlier-rejection guide referenced names
that don't exist in the doc namespace (`masked_mean`, `model`, `ws`); mark them
`<!-- skip: next -->`, matching the convention in docs/guides/orderers.md.

Also correct the `membership_rampup` doctest: with `num_epochs=10, warmup_frac=0.5`
the ramp at epoch 2 is `2 / (0.5 * 9) = 0.444`, not `0.4` (the unit test already
pins 0.4444). Tighten the prose to `warmup_frac * (num_epochs - 1)` to
[…]
