### 2026-07-13 f5e74af
♻️ refactor: rename stream_membership -> posterior_membership, StreamWidthNet -> WidthNet
phasecurvefit is not stream-specific -- streams are one application. Rename the
two stream-flavoured public API names to application-agnostic ones:

  stream_membership -> posterior_membership   (contrasts with the encoder prior)
  StreamWidthNet    -> WidthNet               (matches OrderingNet / TrackNet)

Also renames the private _log_stream_density -> _log_foreground_density and the
TestStreamWidthNet / TestStreamMembership test classes, and lightly generalises
the renamed items' own docstrings (foreg
[…]
