+++ src/phasecurvefit/nn.py
@@ -46,8 +46,8 @@ __all__: tuple[str, ...] = (
     "OrderingTrainingConfig",
     # Membership / outlier rejection (Hogg, Bovy & Lang 2010, sec. 3)
     "MixtureMembershipConfig",
-    "StreamWidthNet",
-    "stream_membership",
+    "WidthNet",
+    "posterior_membership",
     "membership_responsibility",
     "mixture_membership_loss",
     "membership_rampup",
@@ -78,16 +78,16 @@ from ._src.nn import (
     PathAutoencoder,
     RunningMeanDecoder,
     StandardScalerNormalizer,
-    StreamWidthNet,
     TrackNet,
     TrainingConfig,
+    WidthNet,
     encoder_loss,
     fill_ordering_gaps,
     membership_rampup,
     membership_responsibility,
     mixture_membership_loss,
+    posterior_membership,
     sigma_ceiling,
-    stream_membership,
     train_autoencoder,
     train_ordering_net,
     uniform_background_density,
