### 2026-07-13 2a11cf6
✨ feat: membership prob of Hogg, Bovy & Lang (2010) (#43)
Add an opt-in, generative approach to membership / outlier rejection following
Hogg, Bovy & Lang (2010), §3. The default classifier-style membership head
cannot reject outliers: its positive labels come from the orderer (which threads
outliers in), and no gradient connects a point's distance from the fitted track
to its membership. This models the data instead -- each point is drawn either
from the foreground component (a Gaussian of fitted width about the track) or a
flat background -- so membership becomes a calibrated posterior drive
[…]
