+++ src/phasecurvefit/_src/nn/membership.py
@@ -1,0 +1,839 @@
+r"""Mixture-model membership: outlier rejection via a generative likelihood.
+
+Motivation
+----------
+The default membership head is trained as a *classifier*: phase-1
+(`phasecurvefit._src.nn.order_net.encoder_loss`) pushes $p \to 1$ on every
+tracer the orderer visited, and $p \to 0$ on samples drawn uniformly from the
+phase-space bounding box. Two things go wrong.
+
+1. **The positive labels come from the orderer.** An MST (or any spanning
+   structure) has no reject option -- it threads outliers into the ordering
+   along their cheapest edge. Those outliers then land in the positive set, and
+   the network is explicitly supervised to call them members. The membership
+   head becomes a *student of the orderer*, and faithfully reproduces its blind
+   spot.
+
+2. **The reconstruction residual is never used.** The quantity that actually
+   *defines* an outlier -- the distance $r_n = \lVert q_n - x_\theta(\gamma_n)
+   \rVert$ from the fitted track -- is computed by the decoder and then
+   discarded as far as $p$ is concerned. No gradient flows from "this star is
+   far from the track" to "lower its membership probability".
+
+The result is a membership probability that is badly *calibrated*: it saturates
+near 1 for anything remotely near the data manifold, so thresholding it at
+``member_threshold`` separates almost nothing.
+
+The model
+---------
+This module replaces the classifier with a **generative mixture model**, in the
+sense of Hogg, Bovy & Lang (2010), §3 ("Pruning outliers"). Each star is drawn
+either from the stream or from a smooth background:
+
+.. math::
+
+    \mathcal{L}_n = \pi_n \, \mathcal{N}\!\left(q_n \,;\, x_\theta(\gamma_n),\,
+                     \sigma^2(\gamma_n) \mathbb{I}\right)
+                  + (1 - \pi_n) \, \rho_{\mathrm{bg}}
+
+and we minimise the negative log-likelihood $-\sum_n \log \mathcal{L}_n$.
+
+- $\pi_n$ is the encoder's membership output. It plays the role of $(1 - P_b)$
+  in Hogg et al., except that instead of a single global bad-data fraction it is
+  **amortised**: a neural network predicts a per-star mixture weight from the
+  star's own phase-space coordinates.
+- $\mathcal{N}(\cdot)$ is the foreground (stream) density: a Gaussian of width
+  $\sigma(\gamma)$ about the decoded track.
+- $\rho_{\mathrm{bg}}$ is the background density -- a flat density over the
+  field (Hogg et al. use a broad Gaussian $(Y_b, V_b)$; a uniform density is the
+  natural analogue when the "field" is a bounded survey footprint).
+
+Now the gradient with respect to $\pi_n$ is automatic and *correct*: for a star
+close to the track the Gaussian term dominates and $\pi_n \to 1$; for a star far
+from the track the background term dominates and $\pi_n \to 0$. Outlier
+rejection falls out of the likelihood rather than being imposed by a
+self-referential label.
+
+It also replaces the arbitrary ``member_threshold`` with $\rho_{\mathrm{bg}}$,
+which is *physically interpretable* (a background surface density). Hogg et al.
+are blunt about why this matters: sigma-clipping and hand-tuned cuts are "a
+procedure and not the outcome of justifiable modeling".
+
+Why $\sigma$ varies with $\gamma$
+---------------------------------
+A single global $\sigma$ has a breakdown point. If a handful of outliers sit
+only a few stream-widths off the track, the likelihood is happier inflating
[…]
