+++ src/phasecurvefit/nn.py
@@ -44,6 +44,15 @@ __all__: tuple[str, ...] = (
     "train_ordering_net",
     "TrainingConfig",
     "OrderingTrainingConfig",
+    # Membership / outlier rejection (Hogg, Bovy & Lang 2010, sec. 3)
+    "MixtureMembershipConfig",
+    "WidthNet",
+    "posterior_membership",
+    "membership_responsibility",
+    "mixture_membership_loss",
+    "membership_rampup",
+    "sigma_ceiling",
+    "uniform_background_density",
     # Loss functions
     "encoder_loss",
     # Results
@@ -63,6 +72,7 @@ from ._src.nn import (
     AutoencoderResult,
     EncoderExternalDecoder,
     FourierTrackNet,
+    MixtureMembershipConfig,
     OrderingNet,
     OrderingTrainingConfig,
     PathAutoencoder,
@@ -70,8 +80,15 @@ from ._src.nn import (
     StandardScalerNormalizer,
     TrackNet,
     TrainingConfig,
+    WidthNet,
     encoder_loss,
     fill_ordering_gaps,
+    membership_rampup,
+    membership_responsibility,
+    mixture_membership_loss,
+    posterior_membership,
+    sigma_ceiling,
     train_autoencoder,
     train_ordering_net,
+    uniform_background_density,
 )
