### 2026-07-13 5666b97
✨ feat: optional edge-length sigma-clipping for MSTOrderer outlier rejection
The MST threads spatially-isolated interlopers into the ordering along their one
cheapest edge -- it has no reject option. Add an opt-in step to `MSTOrderer` that
rejects them by their MST edge length. `edge_clip_sigma=None` (the default) is an
exact no-op; existing behaviour is unchanged.

Robust, iterated clip on the *spatial* MST edge lengths (not the possibly
velocity-augmented graph weights), in log space -- a multiplicative rule: cut
edges longer than `median * exp(sigma * 1.4826 * MAD(log L))`, never shorter th
[…]
