+++ src/phasecurvefit/_src/orderers/mst.py
@@ -47,6 +47,15 @@ from phasecurvefit._src.custom_types import VectorComponents
 
 OnDisconnected = Literal["raise", "warn", "largest"]
 _TINY = 1e-12
+# An edge must be at least this multiple of the median length to be a clip
+# candidate. Floors the (multiplicative) threshold so a uniformly-sampled
+# backbone -- where the robust spread collapses to ~0 -- is not shredded by
+# microscopic edge-length variation.
+_EDGE_CLIP_MIN_RATIO = 2.0
+# After cutting, a component is an outlier clump (rejected) only if it holds
+# fewer than this fraction of the working points. Larger pieces are kept and
+# reconnected, so cutting a genuine sparse-region edge never discards stream.
+_EDGE_CLIP_SMALL_FRAC = 0.01
 
 
 def _edge_cosine(V: np.ndarray, rows: np.ndarray, cols: np.ndarray) -> np.ndarray:
@@ -91,6 +100,63 @@ def _backbone_on_component(
     return nodes[order_local], backbone_nodes, Cb
 
 
+def _sigma_clip_edges(
+    tree: csr_matrix,
+    P: np.ndarray,
+    nodes: np.ndarray,
+    *,
+    sigma: float,
+    max_iters: int,
+) -> np.ndarray:
+    """Reject outlier nodes by robust, iterated MST edge-length clipping.
+
+    Works on the *spatial* edge lengths (not the possibly velocity-augmented
+    graph weights) in *log* space -- a multiplicative rule, since lengths are
+    positive and heavy-tailed. Each iteration:
+
+    1. cut every edge longer than
+       ``median(L) * exp(sigma * 1.4826 * MAD(log L))``, but never one shorter
+       than ``_EDGE_CLIP_MIN_RATIO * median(L)`` (the floor keeps a
+       uniformly-sampled backbone, where the robust spread is ~0, from being
+       shredded, and still catches a large jump when ``MAD`` is degenerate);
+    2. split the tree at the cut edges and **reject only the small components**
+       (< ``_EDGE_CLIP_SMALL_FRAC`` of the working points) -- the isolated
+       interlopers. Larger pieces are retained and reconnect through the intact
+       ``tree``, so cutting a genuine sparse-region edge cannot discard half the
+       stream;
+    3. recompute the statistic on the survivors and repeat, until nothing small
+       is split off (or ``max_iters``).
+
+    Returns the surviving node set (a subset of ``nodes``, in ascending order).
+    """
+    log_floor = np.log(_EDGE_CLIP_MIN_RATIO)
+    current = nodes
+    for _ in range(max_iters):
+        sub = tree[current][:, current].tocoo()
+        upper = sub.row < sub.col  # undirected edges, once each
+        ei, ej = sub.row[upper], sub.col[upper]
+        if ei.size == 0:
+            break
+        loglen = np.log(np.linalg.norm(P[current[ei]] - P[current[ej]], axis=1))
+        med = float(np.median(loglen))
+        scale = 1.4826 * float(np.median(np.abs(loglen - med)))
+        cut = loglen > med + max(sigma * scale, log_floor)
+        if not cut.any():
+            break
+        m = current.size
+        keep = ~cut
+        g = csr_matrix((np.ones(int(keep.sum())), (ei[keep], ej[keep])), shape=(m, m))
+        g = g.maximum(g.T)
+        _, labels = connected_components(g, directed=False)
+        sizes = np.bincount(labels)
+        size_min = max(2, int(np.ceil(_EDGE_CLIP_SMALL_FRAC * m)))
+        small = np.isin(labels, np.flatnonzero(sizes < size_min))
[…]
