### 2026-07-13 2bdc454
✨ feat: unify ordering under order(); deprecate walk_local_flow (#36)
Make pcf.order(positions, velocities[, orderer]) the primary entry point and
deprecate the free function walk_local_flow (removal in v0.4).

Core (dependency inversion): the real implementation moves to a private,
plum-dispatched _local_flow_walk (plain + Quantity). LocalFlowOrderer.order and
the interop Quantity path now call it directly, and the public walk_local_flow
becomes a thin shim that emits a DeprecationWarning and forwards to it. order()
gains a default orderer (LocalFlowOrderer, via a lazy import to avoid the
bas
[…]
