+++ tests/benchmarks/conftest.py
@@ -113,14 +113,14 @@ def simple_3d_stream(rng_key: PRNGKeyArray) -> tuple[dict, dict]:
 def simple_wlf_result(simple_2d_stream):
     """Create a phase-flow walk result."""
     pos, vel = simple_2d_stream
-    return pcf.walk_local_flow(pos, vel, start_idx=0, metric_scale=1.0)
+    return pcf.order(pos, vel)
 
 
 @pytest.fixture
 def medium_wlf_result(medium_2d_stream):
     """Create a phase-flow walk result for 100 points."""
     pos, vel = medium_2d_stream
-    return pcf.walk_local_flow(pos, vel, start_idx=0, metric_scale=1.0)
+    return pcf.order(pos, vel)
 
 
 @pytest.fixture
@@ -145,7 +145,7 @@ def medium_autoencoder(medium_wlf_result, rng_key):
 def trained_autoencoder(simple_2d_stream, rng_key):
     """Create and train an autoencoder."""
     pos, vel = simple_2d_stream
-    walkresult = pcf.walk_local_flow(pos, vel, start_idx=0, metric_scale=1.0)
+    walkresult = pcf.order(pos, vel)
 
     normalizer = pcf.nn.StandardScalerNormalizer(
         walkresult.positions, walkresult.velocities

[…]
