### 2026-07-13 b2cff1b
Backport PR #45: ✨ feat: unify ordering under `order()`; deprecate `walk_local_flow` (#36)
