+++ tests/test_quantity_support.py
@@ -107,8 +107,10 @@ class TestWalkLocalFlowWithPlainArrays:
 
     def test_walk_local_flow_plain_arrays(self, plain_positions, plain_velocities):
         """Test pcf.walk_local_flow works with plain array inputs."""
-        result = pcf.walk_local_flow(
-            plain_positions, plain_velocities, start_idx=0, metric_scale=0.5
+        result = pcf.order(
+            plain_positions,
+            plain_velocities,
+            pcf.orderers.LocalFlowOrderer(metric_scale=0.5),
         )
 
         # Should visit all 4 points in order
@@ -128,12 +130,10 @@ class TestWalkLocalFlowWithPlainArrays:
         # Create WalkConfig with KDTree strategy (k=3 neighbors)
         config = pcf.WalkConfig(strategy=KDTree(k=3))
 
-        result = pcf.walk_local_flow(
+        result = pcf.order(
             plain_positions,
             plain_velocities,
-            start_idx=0,
-            metric_scale=0.5,
-            config=config,
+            pcf.orderers.LocalFlowOrderer(metric_scale=0.5, config=config),
         )
 
         # Should still visit all 4 points
@@ -151,12 +151,11 @@ class TestWalkLocalFlowWithQuantities:
         """Test pcf.walk_local_flow works with Quantity inputs."""
         lam_quantity = u.Q(0.5, "m")
 
-        result = pcf.walk_local_flow(
+        result = pcf.order(
             quantity_positions,
             quantity_velocities,
-            start_idx=0,
-            metric_scale=lam_quantity,
-            usys=unit_system,
+            pcf.orderers.LocalFlowOrderer(metric_scale=lam_quantity),
+            metadata=pcf.StateMetadata(usys=unit_system),
         )
 
         # Should visit all 4 points in order
@@ -170,12 +169,11 @@ class TestWalkLocalFlowWithQuantities:
         """Test that pcf.walk_local_flow preserves Quantity units in output."""
         lam_quantity = u.Q(0.5, "m")
 
-        result = pcf.walk_local_flow(
+        result = pcf.order(
             quantity_positions,
             quantity_velocities,
-            start_idx=0,
-            metric_scale=lam_quantity,
-            usys=unit_system,
+            pcf.orderers.LocalFlowOrderer(metric_scale=lam_quantity),
+            metadata=pcf.StateMetadata(usys=unit_system),
         )
 
         # Check that positions have units preserved
@@ -203,17 +201,18 @@ class TestWalkLocalFlowWithQuantities:
         lam_quantity = u.Q(0.5, "m")
 
         # Run with plain arrays
-        result_plain = pcf.walk_local_flow(
-            plain_positions, plain_velocities, start_idx=0, metric_scale=lam_plain
+        result_plain = pcf.order(
+            plain_positions,
+            plain_velocities,
+            pcf.orderers.LocalFlowOrderer(metric_scale=lam_plain),
         )
 
         # Run with quantities
-        result_quantity = pcf.walk_local_flow(
+        result_quantity = pcf.order(
             quantity_positions,
             quantity_velocities,
-            start_idx=0,
-            metric_scale=lam_quantity,
-            usys=unit_system,
+            pcf.orderers.LocalFlowOrderer(metric_scale=lam_quantity),
+            metadata=pcf.StateMetadata(usys=unit_system),
         )
 
         # Ordered indices should match
@@ -238,12 +237,11 @@ class TestWalkLocalFlowWithQuantities:
         # Test with higher lambda value (more momentum-dependent)
         lam_quantity = u.Q(2.0, "m")
 
[…]
