+++ tests/usage/test_epitrochoid.py
@@ -122,14 +122,15 @@ def test_epitrochoid_autoencoder_fit() -> plt.Figure:
     max_dist = u.Q(40, "m")
 
     # Perform walk
-    walkresult = pcf.walk_local_flow(
+    walkresult = pcf.order(
         qs,
         ps,
-        start_idx=start_idx,
-        metric_scale=metric_scale,
-        max_dist=max_dist,
-        config=config,
-        direction="forward",
+        pcf.orderers.LocalFlowOrderer(
+            start_idx=start_idx,
+            metric_scale=metric_scale,
+            max_dist=max_dist,
+            config=config,
+        ),
         metadata=pcf.StateMetadata(usys=usys),
     )
 
