+++ tests/test_algorithm.py
@@ -16,7 +16,7 @@ class TestWalkLocalFlowResult:
         q = {"x": jnp.array([1.0, 2.0, 3.0]), "y": jnp.array([1.0, 2.0, 3.0])}
         p = {"x": jnp.array([1.0, 1.0, 1.0]), "y": jnp.array([0.0, 0.0, 0.0])}
 
-        result = pcf.walk_local_flow(q, p, start_idx=0, metric_scale=0.0)
+        result = pcf.order(q, p, pcf.orderers.LocalFlowOrderer(metric_scale=0.0))
 
         # Check that result has the expected attributes (NamedTuple)
         assert hasattr(result, "indices")
@@ -33,7 +33,9 @@ class TestWalkLocalFlowResult:
         q = {"x": jnp.array([3.0, 1.0, 2.0]), "y": jnp.array([30.0, 10.0, 20.0])}
         p = {"x": jnp.array([0.3, 0.1, 0.2]), "y": jnp.array([3.0, 1.0, 2.0])}
 
-        result = pcf.walk_local_flow(q, p, start_idx=1, metric_scale=0.0)
+        result = pcf.order(
+            q, p, pcf.orderers.LocalFlowOrderer(start_idx=1, metric_scale=0.0)
+        )
         ordered_pos, ordered_vel = pcf.order_w(result)
 
         # Order starts from index 1, then finds nearest neighbors
@@ -50,7 +52,7 @@ class TestNearestNeighborsWithMomentum:
         q = {"x": jnp.array([0.0, 1.0, 2.0, 3.0, 4.0])}
         p = {"x": jnp.array([1.0, 1.0, 1.0, 1.0, 1.0])}
 
-        result = pcf.walk_local_flow(q, p, start_idx=0, metric_scale=1.0)
+        result = pcf.order(q, p)
 
         # Should follow the line in order
         assert jnp.array_equal(result.indices, jnp.array([0, 1, 2, 3, 4]))
@@ -62,7 +64,7 @@ class TestNearestNeighborsWithMomentum:
         p = {"x": jnp.array([-1.0, -1.0, -1.0, -1.0, -1.0])}
 
         # Start from middle, should go backward due to velocity
-        result = pcf.walk_local_flow(q, p, start_idx=2, metric_scale=1.0)
+        result = pcf.order(q, p, pcf.orderers.LocalFlowOrderer(start_idx=2))
         # With momentum, should prefer going in velocity direction (backward)
         # First step from 2: velocity points to -x, so prefer 1 over 3
         assert result.indices[1] == 1
@@ -74,13 +76,13 @@ class TestNearestNeighborsWithMomentum:
         p = {"x": jnp.array([1.0, 1.0, 1.0, 1.0, 1.0])}
 
         # Forward walk from leftmost point should go toward higher indices
-        result_forward = pcf.walk_local_flow(q, p, start_idx=0, metric_scale=1.0)
+        result_forward = pcf.order(q, p)
         # First step should be to index 1
         assert result_forward.indices[1] == 1
 
         # Reverse walk from rightmost point should go toward lower indices
-        result_backward = pcf.walk_local_flow(
-            q, p, start_idx=4, metric_scale=1.0, direction="backward"
+        result_backward = pcf.order(
+            q, p, pcf.orderers.LocalFlowOrderer(start_idx=4, direction="backward")
         )
         # First step should be to index 3
         assert result_backward.indices[1] == 3
@@ -91,14 +93,14 @@ class TestNearestNeighborsWithMomentum:
         p = {"x": jnp.array([1.0, 1.0, 1.0, 1.0, 1.0])}
 
         # Using backward parameter
-        result_backward_param = pcf.walk_local_flow(
[…]
