### 2026-07-13 602b6e2
🐛 fix: exclude self from KDTree query neighbors (#46)
The KD-tree query always returns the current point itself (distance 0)
as its nearest neighbor, consuming one of the `k` candidate slots. Since
the current point is already visited, only `k - 1` candidates were
usable, so small `k` could deadlock the walk: at a point whose sole
non-self neighbor is already visited, every candidate is masked and the
walk terminates prematurely (dropping points BruteForce would reach).

Query one extra neighbor and drop the self entry so `k` means `k` usable
candidates. Clamp the query size to the point count
[…]
