### 2026-07-13 1335eea
Backport PR #46: 🐛 fix: KDTree query drops points vs BruteForce; test collection error
