### 2026-07-13 cd26bd5
Backport PR #45: ✨ feat: unify ordering under `order()`; deprecate `walk_local_flow` (#36) (#47)
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
