+++ tests/test_kdtree.py
@@ -23,7 +23,9 @@ def test_walk_local_flow_kdtree_matches_bruteforce(config):
         "y": jnp.array([0.2, 0.2, 0.2, 0.2]),
     }
 
-    res = pcf.walk_local_flow(pos, vel, start_idx=0, metric_scale=0.5, config=config)
+    res = pcf.order(
+        pos, vel, pcf.orderers.LocalFlowOrderer(metric_scale=0.5, config=config)
+    )
     # For this simple dataset, both strategies should visit all points in order
     assert res.all_visited
     assert (res.indices == jnp.array([0, 1, 2, 3])).all()
@@ -40,20 +42,20 @@ def test_walk_local_flow_kdtree_k_parameter():
     }
 
     # Run with KD-tree using small k
-    res_small = pcf.walk_local_flow(
+    res_small = pcf.order(
         pos,
         vel,
-        start_idx=0,
-        metric_scale=0.5,
-        config=pcf.WalkConfig(strategy=pcf.strats.KDTree(k=3)),
+        pcf.orderers.LocalFlowOrderer(
+            metric_scale=0.5, config=pcf.WalkConfig(strategy=pcf.strats.KDTree(k=3))
+        ),
     )
     # Run with KD-tree using larger k
-    res_large = pcf.walk_local_flow(
+    res_large = pcf.order(
         pos,
         vel,
-        start_idx=0,
-        metric_scale=0.5,
-        config=pcf.WalkConfig(strategy=pcf.strats.KDTree(k=8)),
+        pcf.orderers.LocalFlowOrderer(
+            metric_scale=0.5, config=pcf.WalkConfig(strategy=pcf.strats.KDTree(k=8))
+        ),
     )
 
     # Both should produce valid ordered results
