### 2026-08-01 7b51a0e
build(deps): bump the actions group with 4 updates
Bumps the actions group with 4 updates: [pypa/gh-action-pypi-publish](https://github.com/pypa/gh-action-pypi-publish), [astral-sh/setup-uv](https://github.com/astral-sh/setup-uv), [CodSpeedHQ/action](https://github.com/codspeedhq/action) and [actions/labeler](https://github.com/actions/labeler).


Updates `pypa/gh-action-pypi-publish` from 1.14.0 to 1.14.2
- [Release notes](https://github.com/pypa/gh-action-pypi-publish/releases)
- [Commits](https://github.com/pypa/gh-action-pypi-publish/compare/cef221092ed1bacb1cc03d23a2d87d1d172e277b...dc376
[…]
